Data Set with Alaska Moose Counts.

A data set that can be used with the sptotal package. In this example, the counts are of moose on 860 sites of equal area.

AKmoose

Format

A SpatialPolygonsDataFrame object in the format of the (sp) package. The data frame AKmoose@data contains 860 rows and 4 columns:

CENTRLAT

The latitude of the centroid for each site

CENTRLON

The latitude of the centroid for each site

STRAT

A stratification variable

TOTAL

The total moose count on each site

Source

Alaska Department of Fish and Game, Division of Wildlife Conservation has released this data set under the CC0 (creative commons) license. To the extent possible under law, Alaska Department of Fish and Game, Division of Wildlife Conservation waives all copyright and related or neighboring rights to An Alaskan GSPE (Geospatial Population Estimator) Survey of Moose, AKmoose.rda. This work is published from: United States.

Examples

data(AKmoose)
names(AKmoose@data)
#> [1] "elev_mean"   "strat"       "surveyed"    "census_area" "total"      
summary(AKmoose@data)
#>    elev_mean      strat   surveyed census_area     total    
#>  Min.   : 95.12   L:602   0:642    0:773       0      :750  
#>  1st Qu.:136.60   M:258   1:218    1: 87       2      : 21  
#>  Median :176.91                                3      : 21  
#>  Mean   :208.83                                4      : 11  
#>  3rd Qu.:239.03                                7      :  8  
#>  Max.   :656.58                                6      :  7  
#>                                                (Other): 42