Data Set with Alaska Moose Counts.

A data set that can be used with the sptotal package. In this example, the counts are of moose on 860 sites of equal area.

AKmoose_df

Format

A dataframe object. The data frame AKmoose_df contains 860 rows and 4 columns:

elev_mean

The mean elevation for each sitefor each site

strat

A stratification variable (either L or M)

surveyed

Assigned a 1 if the site was surveyed and a 0 otherwise

total

The total moose count on each site (NA if the site was not surveyed)

x

the x-coordinate centroid of the site (TM)

y

the y-coordinate centroid of the site (TM)

lon

the longitudinal centroid of the site

lat

the latitudinal centroid of the site

Source

Alaska Department of Fish and Game, Division of Wildlife Conservation has released this data set under the CC0 (creative commons) license. To the extent possible under law, Alaska Department of Fish and Game, Division of Wildlife Conservation waives all copyright and related or neighboring rights to An Alaskan GSPE (Geospatial Population Estimator) Survey of Moose, AKmoose.rda. This work is published from: United States.

Examples

data(AKmoose_df)
names(AKmoose_df)
#> [1] "elev_mean" "strat"     "surveyed"  "total"     "x"         "y"        
#> [7] "lon"       "lat"      
summary(AKmoose_df)
#>    elev_mean      strat      surveyed          total              x         
#>  Min.   : 95.12   L:602   Min.   :0.0000   Min.   : 0.000   Min.   :  0.00  
#>  1st Qu.:136.60   M:258   1st Qu.:0.0000   1st Qu.: 0.000   1st Qu.: 38.89  
#>  Median :176.91           Median :0.0000   Median : 1.000   Median : 71.90  
#>  Mean   :208.83           Mean   :0.2535   Mean   : 3.404   Mean   : 73.70  
#>  3rd Qu.:239.03           3rd Qu.:1.0000   3rd Qu.: 4.000   3rd Qu.:105.93  
#>  Max.   :656.58           Max.   :1.0000   Max.   :40.000   Max.   :144.72  
#>                                            NA's   :642                      
#>        y               lon              lat       
#>  Min.   :  0.00   Min.   :-148.6   Min.   :62.55  
#>  1st Qu.: 33.51   1st Qu.:-147.9   1st Qu.:62.85  
#>  Median : 56.36   Median :-147.2   Median :63.05  
#>  Mean   : 57.64   Mean   :-147.2   Mean   :63.07  
#>  3rd Qu.: 81.77   3rd Qu.:-146.5   3rd Qu.:63.28  
#>  Max.   :130.40   Max.   :-145.8   Max.   :63.72  
#>