Dissolved Organic Carbon in U.S. Lakes

These data contain dissolved organic carbon (DOC) in National Lakes Data from the U.S. Environmental Protection Agency

USlakes

Format

A data frame with 1206 rows and 9 variables:

XCOORD

x-coordinate from US Contiguous Albers Equal Area Conic projection

YCOORD

y-coordinate from US Contiguous Albers Equal Area Conic projection

DOC_RESULT

Analyte value, in mg/L, for Dissolved Organic Carbon

ELEVATION

Elevation at lake coordinates (LAT_DD_N83, LON_DD_N83) from NHD Digital Elevation Map layer

FCIBIG_LIT

Fish cover: index of fish cover due to large structures in the littoral zone

RVFCGNDBARE_RIP

riparian zone and vegetation: fraction of ground lacking cover in the riparian zone

RVFCGNDWOODY_RIP

riparian zone and vegetation: fraction of ground cover by woody vegetation in the riparian zone

RVFPUNDWOODY_RIP

riparian zone and vegetation: fraction of understory with nonwoody cover present in the riparian zone

UID

A unique lake identifier in the EPA lake survey datebases

Source

National Aquatic Resource Surveys webpage. We combined site data, DOC data, and habitat metrics to create a data set of 1206 lakes in the conterminous United States.

Examples

data(USlakes)
names(USlakes)

#> [1] "UID"              "DOC_RESULT"       "ELEVATION"        "RVFPUNDWOODY_RIP"
#> [5] "FCIBIG_LIT"       "RVFCGNDBARE_RIP"  "RVFCGNDWOODY_RIP" "XCOORD"          
#> [9] "YCOORD"          

summary(USlakes)

#>       UID            DOC_RESULT        ELEVATION       RVFPUNDWOODY_RIP
#>  Min.   :   6160   Min.   :  0.230   Min.   : -53.27   Min.   :0.0000  
#>  1st Qu.:   6967   1st Qu.:  3.300   1st Qu.: 184.24   1st Qu.:0.7000  
#>  Median :   7594   Median :  5.405   Median : 329.25   Median :0.9000  
#>  Mean   : 125183   Mean   :  8.320   Mean   : 631.72   Mean   :0.7677  
#>  3rd Qu.:   8330   3rd Qu.:  8.915   3rd Qu.: 683.02   3rd Qu.:1.0000  
#>  Max.   :1000223   Max.   :515.810   Max.   :3594.97   Max.   :1.0000  
#>    FCIBIG_LIT      RVFCGNDBARE_RIP   RVFCGNDWOODY_RIP      XCOORD        
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000   Min.   :-2293153  
#>  1st Qu.:0.03014   1st Qu.:0.05405   1st Qu.:0.08637   1st Qu.: -884661  
#>  Median :0.09500   Median :0.15921   Median :0.20213   Median :  120741  
#>  Mean   :0.15485   Mean   :0.21033   Mean   :0.23886   Mean   :   84352  
#>  3rd Qu.:0.21250   3rd Qu.:0.31942   3rd Qu.:0.35736   3rd Qu.:  935131  
#>  Max.   :1.26000   Max.   :0.99200   Max.   :0.97793   Max.   : 2240149  
#>      YCOORD        
#>  Min.   :-1270105  
#>  1st Qu.:  116234  
#>  Median :  542813  
#>  Mean   :  472509  
#>  3rd Qu.:  921461  
#>  Max.   : 1522628