Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -1.24979680
#> 2      2    1 -5.62529471
#> 3      3    1 -4.40303666
#> 4      4    1 -2.84947055
#> 5      5    1 -1.96997193
#> 6      6    1 -2.55881604
#> 7      7    1 -4.55898849
#> 8      8    1 -2.37312219
#> 9      9    1 -3.68157898
#> 10    10    1 -0.70856481
#> 11     1    2 -3.75458221
#> 12     2    2 -3.32319808
#> 13     3    2 -5.13095643
#> 14     4    2 -1.80596201
#> 15     5    2 -3.01164962
#> 16     6    2 -1.27197787
#> 17     7    2 -1.57175189
#> 18     8    2 -2.58899302
#> 19     9    2 -0.79967620
#> 20    10    2  0.60016916
#> 21     1    3 -2.80463681
#> 22     2    3 -2.31183053
#> 23     3    3 -4.64561538
#> 24     4    3 -2.56708378
#> 25     5    3 -0.31998733
#> 26     6    3 -1.38954780
#> 27     7    3 -2.94180033
#> 28     8    3 -1.50920608
#> 29     9    3  1.10465691
#> 30    10    3  0.39207730
#> 31     1    4 -3.65304148
#> 32     2    4 -4.08705698
#> 33     3    4 -1.91262941
#> 34     4    4 -1.48203067
#> 35     5    4 -0.73848642
#> 36     6    4 -2.91210886
#> 37     7    4 -3.25755209
#> 38     8    4 -1.33777147
#> 39     9    4 -0.90782343
#> 40    10    4  0.38649320
#> 41     1    5 -0.77559867
#> 42     2    5 -2.70454263
#> 43     3    5 -3.77182823
#> 44     4    5 -3.07013913
#> 45     5    5 -0.14569422
#> 46     6    5 -2.45478348
#> 47     7    5 -0.88321075
#> 48     8    5 -1.42332073
#> 49     9    5  1.05452819
#> 50    10    5 -1.70317005
#> 51     1    6 -1.09544025
#> 52     2    6 -1.04379961
#> 53     3    6 -1.52411249
#> 54     4    6 -1.00316410
#> 55     5    6 -0.05648131
#> 56     6    6 -0.64566306
#> 57     7    6 -1.64171765
#> 58     8    6  0.81262213
#> 59     9    6 -0.09816934
#> 60    10    6 -2.00052041
#> 61     1    7  0.80616823
#> 62     2    7 -1.49195318
#> 63     3    7 -1.48330290
#> 64     4    7 -2.78330010
#> 65     5    7  0.15733000
#> 66     6    7 -0.42561851
#> 67     7    7 -1.26249840
#> 68     8    7  0.08482872
#> 69     9    7 -1.20252927
#> 70    10    7 -0.50460685
#> 71     1    8 -1.17203448
#> 72     2    8 -1.97492450
#> 73     3    8 -1.65512467
#> 74     4    8 -3.00876150
#> 75     5    8  0.37669970
#> 76     6    8 -1.82105323
#> 77     7    8  1.07767823
#> 78     8    8 -1.38162742
#> 79     9    8 -0.67150749
#> 80    10    8  1.92387870
#> 81     1    9  1.59064649
#> 82     2    9 -1.61053566
#> 83     3    9 -3.11755424
#> 84     4    9 -0.03289967
#> 85     5    9 -0.46923387
#> 86     6    9  0.06595792
#> 87     7    9 -0.63858338
#> 88     8    9  0.35011618
#> 89     9    9 -1.32003437
#> 90    10    9  1.72503119
#> 91     1   10 -1.58794651
#> 92     2   10 -1.40173694
#> 93     3   10 -1.95667573
#> 94     4   10  0.16966831
#> 95     5   10 -0.24020129
#> 96     6   10 -1.08781641
#> 97     7   10 -2.59867916
#> 98     8   10 -2.55319847
#> 99     9   10 -2.27544085
#> 100   10   10 -0.41120837