Create a default map from predictions

Creates a default map for the predictions of unobserved sites. Note that all predictions are stored in a data frame in the output of predict.slmfit(). Therefore, if a user would like to create his or her own plot, he or she can easily do so using this data frame.

get.predplot(x)

Arguments

x

the output of the predict.slmfit() function, of class predict.slmfit

Value

a plot with x-coordinates on the x-axis and y-coordinates on the y-axis that is coloured by predictions, with points with an X denoting that a site was sampled and filled circles denoting unsampled sites.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predobj <- predict(slmobj)
## Not Run
## get.predplot(predobj)