Perform Finite Population Block Kriging

Uses an object of class slmfit from the slmfit() function to predict the response on the unsampled sites. The column of the data set that has the response should have numeric values for the observed response on the sampled sites and `NA` for any site that was not sampled. Note that there is no newdata argument to predict.slmfit(): any point in space for which a prediction is needed should be included in the original data set in slmfit() with the response variable as NA.

# S3 method for class 'slmfit'
predict(object, wtscol = NULL, conf_level = 0.9, ...)

Arguments

object

is an object generated from slmfit()

wtscol

is the name of the column that contains the weights for prediction. The default setting predicts the population total

conf_level

is the confidence level for a prediction interval, 0.90 by default

...

further arguments passed to or from other methods.

Value

a list with

the estimated population total

the estimated prediction variance

a data frame containing

x-coordinates

y-coordinates

density predictions

count predictions

site-by-site density prediction variances

site-by-site count prediction variances

indicator variable for whether or not the each site was sampled

estimated mean for each site

area of each site

vector with estimated covariance parameters

the formula used to fit the model in slmfit()

the confidence level

the confidence interval bounds

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predict(slmobj)
#> Prediction Info:
#>        Prediction    SE 90% LB 90% UB
#> counts      813.2 24.64  772.6  853.7
#>        Numb. Sites Sampled Total Numb. Sites Total Observed Average Density
#> counts                  35                40            679           26.77