Prints a short summary for the predict.slmfit() function.

This function uses the object that is output from predict.slmfit().

# S3 method for predict.slmfit
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

x
is a prediction object generated from predict.slmfit()

digits
is the number of digits to be displayed in the model output

...
further arguments passed to or from other methods.

Value

a list with

model formula

summary statistics for the residuals.

a table of fixed effects estimates and associated standard errors.

estimated spatial covariance parameter estimates.

generalized r-squared value.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
print(predict(slmobj))

#> Prediction:
#>       [,1]
#> [1,] 813.2
#> 
#>  Standard Error:
#>       [,1]
#> [1,] 24.64
#> 
#> Prediction Data Frame Preview: First 6 and Last 6 Observations
#>    counts      pred1       pred2 xcoords ycoords   dummyvar areavar xcoordsUTM_
#> 1      24 0.04819231  0.66770938       1       1 0.83624325     0.5           1
#> 2      NA 0.07029807 -0.45115262       2       1 0.05769293     1.0           2
#> 3      21 0.26142376 -0.03514149       3       1 0.43846481     1.0           3
#> 4      11 0.54271551  0.05692399       4       1 0.90608285     0.5           4
#> 5      18 0.99896424  1.27839185       5       1 0.20321972     0.5           5
#> 6      25 0.39053919 -0.78926896       6       1 0.25956914     1.0           6
#> 35     17 0.54991785  1.18865691       7       5 0.82760232     1.0           7
#> 36     33 0.29043608  0.74671108       1       6 0.02721888     1.0           1
#> 37     14 0.52181727  0.37247065       2       6 0.33909108     1.0           2
#> 38     22 0.42858761 -0.56009555       3       6 0.33432158     1.0           3
#> 39     21 0.32678279  0.50311496       4       6 0.52822020     0.5           4
#> 40     23 0.80063605 -0.72036048       5       6 0.27165047     1.0           5
#>    ycoordsUTM_ counts_pred_density counts_pred_count counts_predvar_density
#> 1            1            48.00000          24.00000                 0.0000
#> 2            1            31.01862          31.01862               125.2961
#> 3            1            21.00000          21.00000                 0.0000
#> 4            1            22.00000          11.00000                 0.0000
#> 5            1            36.00000          18.00000                 0.0000
#> 6            1            25.00000          25.00000                 0.0000
#> 35           5            17.00000          17.00000                 0.0000
#> 36           6            33.00000          33.00000                 0.0000
#> 37           6            14.00000          14.00000                 0.0000
#> 38           6            22.00000          22.00000                 0.0000
#> 39           6            42.00000          21.00000                 0.0000
#> 40           6            23.00000          23.00000                 0.0000
#>    counts_predvar_count counts_sampind counts_muhat counts_areas
#> 1                0.0000              1     26.34639          0.5
#> 2              125.2961              0     26.15237          1.0
#> 3                0.0000              1     26.63426          1.0
#> 4                0.0000              1     27.23214          0.5
#> 5                0.0000              1     28.43135          0.5
#> 6                0.0000              1     26.73825          1.0
#> 35               0.0000              1     27.48915          1.0
#> 36               0.0000              1     26.86122          1.0
#> 37               0.0000              1     27.25672          1.0
#> 38               0.0000              1     26.86550          1.0
#> 39               0.0000              1     26.88379          0.5
#> 40               0.0000              1     27.59593          1.0