Prints a short summary for the predict.stratafit() function.

This function uses the object that is output from predict.stratafit() of class predict.stratafit..

# S3 method for predict.stratafit
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

x
is a prediction object generated from predict.stratafit()

digits
is the number of digits to be displayed in the model output

...
further arguments passed to or from other methods.

Value

a list with

a table of predictions, standard errors, and confidence intervals for each stratum and for the total.

Examples

data(exampledataset) ## load a toy data set
exampledataset$strata <- c(rep("A", 19), rep("B", 21))
strataobj <- stratafit(formula = counts ~ pred1 + pred2,
 data = exampledataset, stratacol = "strata",
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predict(strataobj)

#> 
#> Prediction and Confidence Intervals:
#>       Prediction     SE 90% LB 90% UB
#> A          364.9  9.213  349.8  380.1
#> B          412.2  9.889  396.0  428.5
#> Total      777.1 13.515  754.9  799.4