Prints the fitted coefficient table of a fitted spatial linear model.

This function uses the object that is output from slmfit() of class slmfit.

# S3 method for slmfit
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

x
is an object generated from slmfit()

digits
is the number of digits to be displayed in the model output

...
further arguments passed to or from other methods.

Value

a list with

model formula

summary statistics for the residuals.

a table of fixed effects estimates and associated standard errors.

estimated spatial covariance parameter estimates.

generalized r-squared value.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
print(slmobj)

#> 
#> Coefficients:
#> (Intercept)       pred1       pred2 
#>      26.104       2.055       0.214 
#> 
#> Covariance Parameters:
#>       Nugget Partial Sill        Range 
#>    2.323e-04    1.544e+02    8.240e-01