Prints the summary of a fitted spatial linear model.

This function uses the object that is output from summary.slmfit().

# S3 method for summary.slmfit
print(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

Arguments

x

is an summary object generated from summary.slmfit()

digits

is the number of digits to be displayed in the model output

signif.stars

is an option to show which predictors are significant.

...

further arguments passed to or from other methods.

Value

a list with

model formula

summary statistics for the residuals.

a table of fixed effects estimates and associated standard errors.

estimated spatial covariance parameter estimates.

generalized r-squared value.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
print(summary(slmobj))
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -16.086  -9.120  -4.596   4.963  28.594 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   26.104      4.465   5.847   <2e-16 ***
#> pred1          2.055      6.065   0.339    0.737    
#> pred2          0.214      1.913   0.112    0.912    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            2.323211e-04
#> Partial Sill      1.543624e+02
#> Range             8.240060e-01
#> 
#> Generalized R-squared: 0.00413886