Simulated Spatially Autocorrelated Data.

A simulated data set that can be used with the sptotal package.

simdata

Format

A data frame object including:

x

The x-coordinate for each site

y

The y-coordinate for each site

X1

Simulated independent variable to be used as a predictor

X2

Simulated independent variable to be used as a predictor

X3

Simulated independent variable to be used as a predictor

X4

Simulated independent variable to be used as a predictor

X5

Simulated independent variable to be used as a predictor

X6

Simulated spatially correlated random variable to be used as a predictor

X7

Simulated spatially correlated random variable to be used as a predictor

F1

Simulated factor variable to be used as a predictor

F2

Simulated factor variable to be used as a predictor

Z

The simulated response variable.

wts1

Prediction weights if estimating an overall mean

wts2

Prediction weights for estimating a total over a subset of 25 contiguous plots

Examples

data(simdata)
names(simdata)
#>  [1] "x"    "y"    "X1"   "X2"   "X3"   "X4"   "X5"   "X6"   "X7"   "F1"  
#> [11] "F2"   "Z"    "wts1" "wts2"
summary(simdata)
#>        x                y                X1                 X2          
#>  Min.   :0.0250   Min.   :0.0250   Min.   :-2.99286   Min.   :-2.84025  
#>  1st Qu.:0.2625   1st Qu.:0.2625   1st Qu.:-0.65748   1st Qu.:-0.58639  
#>  Median :0.5000   Median :0.5000   Median : 0.03178   Median : 0.07411  
#>  Mean   :0.5000   Mean   :0.5000   Mean   :-0.01416   Mean   : 0.08079  
#>  3rd Qu.:0.7375   3rd Qu.:0.7375   3rd Qu.: 0.58910   3rd Qu.: 0.74515  
#>  Max.   :0.9750   Max.   :0.9750   Max.   : 3.40043   Max.   : 2.96393  
#>        X3                 X4                  X5                X6         
#>  Min.   :-2.67599   Min.   :-2.252048   Min.   :-3.2961   Min.   :-2.2618  
#>  1st Qu.:-0.68008   1st Qu.:-0.762883   1st Qu.:-0.7519   1st Qu.:-0.9030  
#>  Median :-0.08512   Median :-0.002231   Median :-0.1421   Median :-0.1350  
#>  Mean   :-0.04361   Mean   : 0.007952   Mean   :-0.1336   Mean   :-0.1351  
#>  3rd Qu.: 0.63493   3rd Qu.: 0.696168   3rd Qu.: 0.5012   3rd Qu.: 0.4743  
#>  Max.   : 3.31052   Max.   : 2.533708   Max.   : 2.5289   Max.   : 2.8975  
#>        X7          F1      F2           Z               wts1       
#>  Min.   :-1.9833   1:134   1:80   Min.   : 8.063   Min.   :0.0025  
#>  1st Qu.:-0.2879   2:133   2:80   1st Qu.:11.047   1st Qu.:0.0025  
#>  Median : 0.2629   3:133   3:80   Median :12.101   Median :0.0025  
#>  Mean   : 0.2519           4:80   Mean   :12.086   Mean   :0.0025  
#>  3rd Qu.: 0.7889           5:80   3rd Qu.:13.089   3rd Qu.:0.0025  
#>  Max.   : 2.6919                  Max.   :15.944   Max.   :0.0025  
#>       wts2       
#>  Min.   :0.0000  
#>  1st Qu.:0.0000  
#>  Median :0.0000  
#>  Mean   :0.0625  
#>  3rd Qu.:0.0000  
#>  Max.   :1.0000