Fits a Spatial Linear Model

Estimates regression coefficients and spatial autocorrelation parameters, given spatial coordinates and a model formula.

slmfit(
  formula,
  data,
  xcoordcol,
  ycoordcol,
  areacol = NULL,
  CorModel = "Exponential",
  estmethod = "REML",
  covestimates = c(NA, NA, NA)
)

Arguments

formula
is an R linear model formula specifying the response variable as well as covariates for predicting the response on the unsampled sites.

data
is the data set with the response column, the covariates to be used for the block kriging, and the spatial coordinates for all of the sites.

xcoordcol
is the name of the column in the data frame with x coordinates or longitudinal coordinates

ycoordcol
is the name of the column in the data frame with y coordinates or latitudinal coordinates

areacol
is the name of the column with the areas of the sites. By default, we assume that all sites have equal area, in which case a vector of 1's is used as the areas.

CorModel
is the covariance structure. By default, CorModel is Exponential but other options include the Spherical and Gaussian.

estmethod
is either the default "REML" for restricted maximum likelihood to estimate the covariance parameters and regression coefficients or "ML" to estimate the covariance parameters and regression coefficients. This argument can also be set to "None", in which case covestimates must be provided.

covestimates
is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and estmethod = "None", the the provided vector are treated as the estimators to create the covariance structure.

Value

a list of class slmfit with

the spatial covariance estimates

the regression coefficient estimates

the covariance matrix of the fixed effects

minus two times the log-likeihood of the model

the names of the predictors

the sample size

the name of the covariance model used

a vector of residuals

the design matrix

a vector of the sampled densities

a list containing

formula, the model formula

data, the data set input as the data argument

xcoordcol, the name of the x-coordinate column

ycoordcol, the name of the y-coordinate column

estmethod, either REML or ML

CorModel, the correlation model used

estimated covariance matrix of all sites

Inverted covariance matrix on the sampled sites

the vector of areas.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
summary(slmobj)

#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -16.086  -9.120  -4.596   4.963  28.594 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  26.1044     0.3594  72.642  < 2e-16 ***
#> pred1         2.0554     0.4882   4.211  0.00019 ***
#> pred2         0.2140     0.1540   1.390  0.17411    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            2.323211e-04
#> Partial Sill      1.543624e+02
#> Range             8.240060e-01
#> 
#> Generalized R-squared: 0.00413886