Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2            z
#> 1      1    1 -1.497763175
#> 2      2    1  3.127236739
#> 3      3    1 -2.125117065
#> 4      4    1 -1.259867572
#> 5      5    1 -1.158147536
#> 6      6    1 -1.207776572
#> 7      7    1 -1.534537768
#> 8      8    1 -0.220613898
#> 9      9    1  0.903237585
#> 10    10    1 -0.989020789
#> 11     1    2 -1.161409211
#> 12     2    2  0.559246968
#> 13     3    2  1.427998781
#> 14     4    2 -1.065014216
#> 15     5    2 -2.929466820
#> 16     6    2 -0.738958202
#> 17     7    2 -1.705666399
#> 18     8    2  0.446715818
#> 19     9    2  1.578821530
#> 20    10    2  0.824928842
#> 21     1    3 -1.546813885
#> 22     2    3  1.830922780
#> 23     3    3  0.103635043
#> 24     4    3  0.318522353
#> 25     5    3 -1.156991343
#> 26     6    3  0.201298530
#> 27     7    3 -2.219655618
#> 28     8    3 -1.021626311
#> 29     9    3 -0.015305712
#> 30    10    3 -0.055746635
#> 31     1    4  0.676339979
#> 32     2    4 -0.145445722
#> 33     3    4 -0.595492371
#> 34     4    4 -0.247895128
#> 35     5    4 -0.027144975
#> 36     6    4  1.092712600
#> 37     7    4  1.184124739
#> 38     8    4 -1.452729613
#> 39     9    4 -2.141860588
#> 40    10    4  0.081030808
#> 41     1    5 -1.342725191
#> 42     2    5 -0.221981462
#> 43     3    5  0.640106599
#> 44     4    5 -1.236398257
#> 45     5    5  0.024890790
#> 46     6    5 -1.215742763
#> 47     7    5 -3.571748497
#> 48     8    5 -1.864933712
#> 49     9    5 -1.721662588
#> 50    10    5 -0.669409134
#> 51     1    6 -1.778942497
#> 52     2    6 -0.704221851
#> 53     3    6 -1.452308317
#> 54     4    6 -1.281777175
#> 55     5    6 -1.624702110
#> 56     6    6 -1.110400715
#> 57     7    6 -0.073905661
#> 58     8    6 -1.842743781
#> 59     9    6 -1.100661041
#> 60    10    6 -0.486424946
#> 61     1    7 -1.248203838
#> 62     2    7  0.390137070
#> 63     3    7 -0.498579544
#> 64     4    7  0.002059676
#> 65     5    7 -0.879843366
#> 66     6    7 -0.792001370
#> 67     7    7 -1.507189335
#> 68     8    7 -1.156039129
#> 69     9    7 -2.165987895
#> 70    10    7 -2.087160212
#> 71     1    8 -0.902626782
#> 72     2    8 -0.032964391
#> 73     3    8 -0.574525675
#> 74     4    8 -2.280094484
#> 75     5    8  0.387723813
#> 76     6    8 -0.126637975
#> 77     7    8 -1.772685903
#> 78     8    8 -2.597935021
#> 79     9    8 -1.933401485
#> 80    10    8 -1.271416399
#> 81     1    9 -1.307555970
#> 82     2    9  0.034658728
#> 83     3    9 -1.362719189
#> 84     4    9 -1.573111532
#> 85     5    9  0.237616167
#> 86     6    9 -0.752404946
#> 87     7    9 -1.742561684
#> 88     8    9 -1.555630694
#> 89     9    9  0.264112877
#> 90    10    9 -1.773731046
#> 91     1   10 -1.437403254
#> 92     2   10 -1.740087409
#> 93     3   10 -1.335361328
#> 94     4   10 -1.481973476
#> 95     5   10 -0.069260326
#> 96     6   10  1.452746420
#> 97     7   10 -0.742183256
#> 98     8   10 -1.681835660
#> 99     9   10 -0.636415613
#> 100   10   10 -1.492459718