Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -3.73530999
#> 2      2    1 -3.85254061
#> 3      3    1 -2.37477477
#> 4      4    1 -1.76632348
#> 5      5    1 -0.77198342
#> 6      6    1 -0.87149090
#> 7      7    1  0.46670303
#> 8      8    1 -3.87890666
#> 9      9    1 -0.36606462
#> 10    10    1  0.98429155
#> 11     1    2 -2.85388938
#> 12     2    2 -0.22258685
#> 13     3    2 -2.53244874
#> 14     4    2  0.09588543
#> 15     5    2 -2.46015935
#> 16     6    2 -1.95878853
#> 17     7    2 -0.82325405
#> 18     8    2 -2.86072465
#> 19     9    2 -0.17521818
#> 20    10    2 -1.54466922
#> 21     1    3 -2.14159385
#> 22     2    3 -4.61275259
#> 23     3    3 -1.16149535
#> 24     4    3 -0.72075812
#> 25     5    3 -1.02747231
#> 26     6    3 -2.00458699
#> 27     7    3 -1.91284168
#> 28     8    3 -1.27458529
#> 29     9    3 -0.24181428
#> 30    10    3  0.50190619
#> 31     1    4 -3.68818341
#> 32     2    4 -2.94497702
#> 33     3    4 -2.49477654
#> 34     4    4 -1.63813190
#> 35     5    4 -0.88370762
#> 36     6    4 -4.19997212
#> 37     7    4 -2.98291715
#> 38     8    4  1.03893994
#> 39     9    4 -1.31843114
#> 40    10    4 -0.76356693
#> 41     1    5 -3.64074654
#> 42     2    5 -2.63742902
#> 43     3    5 -3.40688034
#> 44     4    5 -3.69518025
#> 45     5    5 -2.86691955
#> 46     6    5 -5.01731005
#> 47     7    5 -2.45182212
#> 48     8    5  0.07904426
#> 49     9    5 -2.49086858
#> 50    10    5 -2.14135601
#> 51     1    6 -0.78974637
#> 52     2    6 -2.25435974
#> 53     3    6 -1.13330423
#> 54     4    6 -1.34226999
#> 55     5    6 -2.35349153
#> 56     6    6 -3.35969881
#> 57     7    6 -2.04760607
#> 58     8    6 -2.90181425
#> 59     9    6 -1.22536316
#> 60    10    6 -1.30524496
#> 61     1    7 -2.73255385
#> 62     2    7 -2.79024881
#> 63     3    7 -2.94578644
#> 64     4    7 -1.80038428
#> 65     5    7 -0.95624546
#> 66     6    7 -2.38313671
#> 67     7    7 -0.30937942
#> 68     8    7 -1.36150360
#> 69     9    7 -1.46848546
#> 70    10    7 -1.00937325
#> 71     1    8 -2.09774604
#> 72     2    8 -2.72569460
#> 73     3    8 -4.07002709
#> 74     4    8 -2.40957784
#> 75     5    8 -1.01297729
#> 76     6    8 -1.88812332
#> 77     7    8 -0.93706184
#> 78     8    8 -0.85194147
#> 79     9    8 -0.39054726
#> 80    10    8 -0.42048141
#> 81     1    9 -2.53827702
#> 82     2    9 -3.20525919
#> 83     3    9 -1.16653110
#> 84     4    9 -2.66785633
#> 85     5    9 -2.37896338
#> 86     6    9 -0.82839386
#> 87     7    9 -0.46739705
#> 88     8    9  0.05214108
#> 89     9    9 -0.32085143
#> 90    10    9  0.30408178
#> 91     1   10 -2.21191505
#> 92     2   10 -0.65820641
#> 93     3   10 -2.95540645
#> 94     4   10 -2.32935299
#> 95     5   10 -1.46760733
#> 96     6   10 -2.14377410
#> 97     7   10 -1.56059279
#> 98     8   10 -1.65309293
#> 99     9   10  0.59295267
#> 100   10   10 -1.58548676