Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -0.65172768
#> 2      2    1 -2.56972991
#> 3      3    1 -0.39888041
#> 4      4    1 -2.58943265
#> 5      5    1  0.64966702
#> 6      6    1 -1.04651845
#> 7      7    1 -1.30250482
#> 8      8    1 -2.52879442
#> 9      9    1 -1.41530525
#> 10    10    1 -0.63723864
#> 11     1    2 -0.24132766
#> 12     2    2 -0.62425624
#> 13     3    2  0.61016924
#> 14     4    2 -0.55788171
#> 15     5    2  0.43852938
#> 16     6    2 -1.65600151
#> 17     7    2 -2.06098551
#> 18     8    2  0.75000190
#> 19     9    2  2.28438798
#> 20    10    2 -3.49668289
#> 21     1    3  0.30026768
#> 22     2    3 -0.25489998
#> 23     3    3  2.91290889
#> 24     4    3  2.41983105
#> 25     5    3 -0.68212027
#> 26     6    3 -1.81751755
#> 27     7    3  0.27777472
#> 28     8    3 -1.69212240
#> 29     9    3  0.81862360
#> 30    10    3 -1.28213897
#> 31     1    4 -3.23766622
#> 32     2    4  0.40830243
#> 33     3    4  1.20862449
#> 34     4    4  1.49588281
#> 35     5    4  1.41636889
#> 36     6    4 -0.21341015
#> 37     7    4  0.69786801
#> 38     8    4 -2.77010674
#> 39     9    4  0.13659277
#> 40    10    4 -0.31498534
#> 41     1    5  1.16963126
#> 42     2    5  1.32913123
#> 43     3    5  2.12672454
#> 44     4    5  0.98053991
#> 45     5    5 -1.83887739
#> 46     6    5 -1.48576359
#> 47     7    5 -4.44997145
#> 48     8    5 -1.23206228
#> 49     9    5  1.20599665
#> 50    10    5  0.24129266
#> 51     1    6  1.43627493
#> 52     2    6 -0.22731448
#> 53     3    6  0.78627380
#> 54     4    6 -1.15257425
#> 55     5    6 -2.13633479
#> 56     6    6 -1.51534969
#> 57     7    6 -3.37074410
#> 58     8    6 -0.45182262
#> 59     9    6 -1.40984988
#> 60    10    6  0.10977584
#> 61     1    7  0.96378121
#> 62     2    7  2.59472733
#> 63     3    7  2.13323810
#> 64     4    7  0.04076989
#> 65     5    7  0.02748193
#> 66     6    7 -1.28272531
#> 67     7    7 -0.54892257
#> 68     8    7  0.01698643
#> 69     9    7  0.19234996
#> 70    10    7 -0.94941668
#> 71     1    8  1.72423289
#> 72     2    8 -0.45791531
#> 73     3    8 -0.11777679
#> 74     4    8 -0.84655692
#> 75     5    8  0.20902165
#> 76     6    8 -0.90323178
#> 77     7    8 -2.11237063
#> 78     8    8 -2.81034678
#> 79     9    8 -1.44036515
#> 80    10    8 -0.04884892
#> 81     1    9  0.59951906
#> 82     2    9  1.24878261
#> 83     3    9 -0.84223490
#> 84     4    9 -2.91110445
#> 85     5    9 -1.21300835
#> 86     6    9 -0.13294485
#> 87     7    9 -1.29938272
#> 88     8    9 -0.56240174
#> 89     9    9 -1.19693160
#> 90    10    9 -0.07403203
#> 91     1   10  0.47636781
#> 92     2   10  0.54569290
#> 93     3   10 -0.93938014
#> 94     4   10 -2.67358193
#> 95     5   10 -1.02936553
#> 96     6   10 -1.87189759
#> 97     7   10 -2.81290211
#> 98     8   10 -1.83075237
#> 99     9   10  0.80275622
#> 100   10   10 -0.68448148