Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -2.72327954
#> 2      2    1 -0.56401873
#> 3      3    1 -1.94561965
#> 4      4    1 -2.07702608
#> 5      5    1  0.15460128
#> 6      6    1  1.51394032
#> 7      7    1  1.48928947
#> 8      8    1  1.09527123
#> 9      9    1  0.68215617
#> 10    10    1 -1.89232547
#> 11     1    2 -1.28429433
#> 12     2    2 -1.42845093
#> 13     3    2  0.49403885
#> 14     4    2 -0.71616916
#> 15     5    2  0.17882479
#> 16     6    2  0.53389711
#> 17     7    2 -0.37053487
#> 18     8    2  0.45374837
#> 19     9    2 -1.14556146
#> 20    10    2  0.70945211
#> 21     1    3 -4.26191468
#> 22     2    3 -4.92996868
#> 23     3    3 -0.92481530
#> 24     4    3 -0.78752641
#> 25     5    3 -1.45897806
#> 26     6    3  1.51890358
#> 27     7    3 -0.68610115
#> 28     8    3 -0.88609642
#> 29     9    3 -0.51039633
#> 30    10    3  0.62879793
#> 31     1    4 -1.17295824
#> 32     2    4  0.25296036
#> 33     3    4 -3.38766332
#> 34     4    4  0.03709895
#> 35     5    4 -1.04480105
#> 36     6    4 -1.10030333
#> 37     7    4 -0.89794959
#> 38     8    4 -0.74811491
#> 39     9    4 -1.38087154
#> 40    10    4 -1.48595061
#> 41     1    5 -2.78633402
#> 42     2    5 -0.97164970
#> 43     3    5 -3.67004753
#> 44     4    5 -0.88892290
#> 45     5    5 -1.54100874
#> 46     6    5 -1.57757736
#> 47     7    5 -3.56679096
#> 48     8    5 -0.65820543
#> 49     9    5  0.51066012
#> 50    10    5 -0.99560513
#> 51     1    6 -2.32295904
#> 52     2    6 -1.32951563
#> 53     3    6 -3.12752851
#> 54     4    6 -2.15790706
#> 55     5    6  0.41704132
#> 56     6    6 -0.30911955
#> 57     7    6 -0.90063479
#> 58     8    6  1.39932966
#> 59     9    6 -0.24440065
#> 60    10    6  1.60369989
#> 61     1    7 -0.27769569
#> 62     2    7 -0.23492223
#> 63     3    7 -1.38432258
#> 64     4    7 -0.56691097
#> 65     5    7 -0.52777971
#> 66     6    7 -0.88822369
#> 67     7    7  0.37056608
#> 68     8    7  0.80968466
#> 69     9    7  1.14307026
#> 70    10    7  1.85342164
#> 71     1    8 -3.25987977
#> 72     2    8  1.76072076
#> 73     3    8 -1.68277976
#> 74     4    8 -3.72536484
#> 75     5    8 -2.44935330
#> 76     6    8  1.44556518
#> 77     7    8  0.51108153
#> 78     8    8  1.42781033
#> 79     9    8  1.79898370
#> 80    10    8  1.69360559
#> 81     1    9 -2.35790185
#> 82     2    9 -2.58778828
#> 83     3    9 -1.78198409
#> 84     4    9 -1.40245978
#> 85     5    9 -0.02424009
#> 86     6    9 -0.52275801
#> 87     7    9 -2.02865789
#> 88     8    9  0.42222178
#> 89     9    9 -1.44827561
#> 90    10    9  1.32344410
#> 91     1   10 -0.61652303
#> 92     2   10 -2.23410554
#> 93     3   10 -2.14074114
#> 94     4   10 -0.20678701
#> 95     5   10 -0.93040854
#> 96     6   10 -0.17453146
#> 97     7   10  1.87643492
#> 98     8   10  0.64094501
#> 99     9   10  0.82704979
#> 100   10   10  0.26064382