Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -1.14326418
#> 2      2    1 -1.95430269
#> 3      3    1 -2.44889394
#> 4      4    1 -1.07391570
#> 5      5    1  1.41696009
#> 6      6    1  0.97621238
#> 7      7    1  0.45206754
#> 8      8    1  1.52239834
#> 9      9    1  1.52627311
#> 10    10    1  1.07171806
#> 11     1    2 -2.20992651
#> 12     2    2  2.29909745
#> 13     3    2  0.41526424
#> 14     4    2 -1.43060568
#> 15     5    2 -2.58258046
#> 16     6    2 -1.05376482
#> 17     7    2  0.43788760
#> 18     8    2  2.49540640
#> 19     9    2  0.75222453
#> 20    10    2  1.19154223
#> 21     1    3 -1.78971096
#> 22     2    3  0.73494086
#> 23     3    3 -2.07883008
#> 24     4    3 -0.22674722
#> 25     5    3  0.38607153
#> 26     6    3  0.07321229
#> 27     7    3  1.51055609
#> 28     8    3  0.43168729
#> 29     9    3  1.81038071
#> 30    10    3  3.91164187
#> 31     1    4 -0.86568967
#> 32     2    4 -0.52339796
#> 33     3    4 -2.05893078
#> 34     4    4  0.46841328
#> 35     5    4 -0.70377084
#> 36     6    4  1.15044429
#> 37     7    4  1.18921288
#> 38     8    4  0.50591714
#> 39     9    4  1.85406818
#> 40    10    4  2.67151809
#> 41     1    5 -0.15138300
#> 42     2    5 -0.38771954
#> 43     3    5  0.76379126
#> 44     4    5  0.85231108
#> 45     5    5  2.79609671
#> 46     6    5  2.50210743
#> 47     7    5  1.54099391
#> 48     8    5  1.14717357
#> 49     9    5  3.67621796
#> 50    10    5  0.19744830
#> 51     1    6  1.32607010
#> 52     2    6 -0.47102726
#> 53     3    6  0.20722375
#> 54     4    6  1.97458940
#> 55     5    6  0.57735065
#> 56     6    6  1.68567044
#> 57     7    6 -1.22348374
#> 58     8    6 -0.11672058
#> 59     9    6  2.83183787
#> 60    10    6  3.93138918
#> 61     1    7  0.76483407
#> 62     2    7  1.69577795
#> 63     3    7  0.50556415
#> 64     4    7  0.50935204
#> 65     5    7  3.86611681
#> 66     6    7 -0.50896471
#> 67     7    7 -0.11070027
#> 68     8    7  0.45632501
#> 69     9    7  2.23737997
#> 70    10    7  2.25350444
#> 71     1    8 -1.71809119
#> 72     2    8 -0.75949203
#> 73     3    8  0.52779590
#> 74     4    8 -0.62583461
#> 75     5    8 -0.67244107
#> 76     6    8  1.27366591
#> 77     7    8 -0.82091847
#> 78     8    8  2.07507516
#> 79     9    8  2.24914899
#> 80    10    8  0.27873290
#> 81     1    9 -4.36245660
#> 82     2    9 -1.32603727
#> 83     3    9 -2.85263508
#> 84     4    9 -2.06012992
#> 85     5    9 -0.30317037
#> 86     6    9 -0.78508921
#> 87     7    9 -0.51520756
#> 88     8    9 -2.74513709
#> 89     9    9 -0.11227586
#> 90    10    9 -0.20544596
#> 91     1   10 -1.57242541
#> 92     2   10 -2.25174444
#> 93     3   10 -3.12254790
#> 94     4   10 -3.55448243
#> 95     5   10 -2.28183883
#> 96     6   10  0.91769098
#> 97     7   10 -1.81150081
#> 98     8   10  0.89421594
#> 99     9   10 -0.32179676
#> 100   10   10 -1.07816668