Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2            z
#> 1      1    1  0.442993850
#> 2      2    1 -1.180308503
#> 3      3    1 -0.842866648
#> 4      4    1  0.618302859
#> 5      5    1  0.826907018
#> 6      6    1  0.392022920
#> 7      7    1  2.901041035
#> 8      8    1  2.270215929
#> 9      9    1  0.434447621
#> 10    10    1  0.110552581
#> 11     1    2 -1.345997289
#> 12     2    2  0.152897198
#> 13     3    2  0.143280902
#> 14     4    2  0.009941237
#> 15     5    2  1.823313396
#> 16     6    2 -1.376609926
#> 17     7    2  1.857944164
#> 18     8    2  0.755522479
#> 19     9    2  2.356397875
#> 20    10    2  3.341994222
#> 21     1    3  1.033086147
#> 22     2    3  0.422618999
#> 23     3    3  1.083432984
#> 24     4    3  2.268853453
#> 25     5    3  1.775536853
#> 26     6    3  3.309414211
#> 27     7    3  2.955052017
#> 28     8    3 -0.328572901
#> 29     9    3  0.069825214
#> 30    10    3  2.344306569
#> 31     1    4  1.499684437
#> 32     2    4  0.316948431
#> 33     3    4  0.961648781
#> 34     4    4  0.210223388
#> 35     5    4  1.945164870
#> 36     6    4  2.224023721
#> 37     7    4  0.744460432
#> 38     8    4  0.342432793
#> 39     9    4  0.936951176
#> 40    10    4  1.425939118
#> 41     1    5  0.151798272
#> 42     2    5  1.297926798
#> 43     3    5  1.553994368
#> 44     4    5  0.408847948
#> 45     5    5  0.078031686
#> 46     6    5 -0.085129858
#> 47     7    5  0.724637991
#> 48     8    5  2.385552761
#> 49     9    5  3.554615953
#> 50    10    5  2.397836604
#> 51     1    6  2.583961132
#> 52     2    6 -1.795965248
#> 53     3    6 -0.411928825
#> 54     4    6  0.883121749
#> 55     5    6 -0.367623992
#> 56     6    6  2.575224968
#> 57     7    6  0.278292354
#> 58     8    6  1.492416084
#> 59     9    6  1.667350459
#> 60    10    6  2.452985686
#> 61     1    7 -0.458921046
#> 62     2    7  2.568105727
#> 63     3    7 -0.335818599
#> 64     4    7  1.340362200
#> 65     5    7  2.390547979
#> 66     6    7  0.858182270
#> 67     7    7  0.949974836
#> 68     8    7  1.252637597
#> 69     9    7  0.718143936
#> 70    10    7  1.768945947
#> 71     1    8 -0.623368188
#> 72     2    8  2.666424301
#> 73     3    8 -0.157379568
#> 74     4    8  0.373873930
#> 75     5    8 -0.136073576
#> 76     6    8  2.511993533
#> 77     7    8  2.257514096
#> 78     8    8  0.910187883
#> 79     9    8  1.966341133
#> 80    10    8  3.929413069
#> 81     1    9 -0.202451539
#> 82     2    9 -1.008623207
#> 83     3    9 -0.405617366
#> 84     4    9  3.020710840
#> 85     5    9  0.916730855
#> 86     6    9  1.207739138
#> 87     7    9  1.691535422
#> 88     8    9  1.979313779
#> 89     9    9  0.927380623
#> 90    10    9 -0.680401885
#> 91     1   10  0.826926280
#> 92     2   10 -0.332520312
#> 93     3   10 -0.856578048
#> 94     4   10  1.903004776
#> 95     5   10  1.768826756
#> 96     6   10  0.437861527
#> 97     7   10  1.471004822
#> 98     8   10  0.668527575
#> 99     9   10 -0.398981913
#> 100   10   10  3.482162447