Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2            z
#> 1      1    1 -1.588393237
#> 2      2    1 -0.883967826
#> 3      3    1  0.894590655
#> 4      4    1 -0.371038713
#> 5      5    1  1.937446781
#> 6      6    1 -0.512858294
#> 7      7    1  0.743216327
#> 8      8    1  0.460541570
#> 9      9    1 -0.478918906
#> 10    10    1 -0.866508977
#> 11     1    2  2.070427172
#> 12     2    2  1.545190012
#> 13     3    2 -0.877754133
#> 14     4    2  0.316626307
#> 15     5    2 -0.316820251
#> 16     6    2  0.716464393
#> 17     7    2  0.429682896
#> 18     8    2 -0.595063087
#> 19     9    2 -2.140723608
#> 20    10    2 -1.773436560
#> 21     1    3  0.789886498
#> 22     2    3  1.880898544
#> 23     3    3  1.915684784
#> 24     4    3  0.238606254
#> 25     5    3  0.665580056
#> 26     6    3  0.656086091
#> 27     7    3 -0.297033470
#> 28     8    3  0.078740279
#> 29     9    3  0.402535763
#> 30    10    3  0.602308679
#> 31     1    4  0.722963100
#> 32     2    4  0.315001136
#> 33     3    4  2.138250066
#> 34     4    4  2.209903683
#> 35     5    4  0.846401710
#> 36     6    4  0.556596206
#> 37     7    4 -0.085423928
#> 38     8    4  0.360679205
#> 39     9    4  1.503494413
#> 40    10    4  0.047748121
#> 41     1    5  0.274450179
#> 42     2    5  1.410939559
#> 43     3    5  1.802296427
#> 44     4    5  0.296919612
#> 45     5    5  1.399890711
#> 46     6    5  0.888029869
#> 47     7    5 -1.647318222
#> 48     8    5 -1.574428624
#> 49     9    5 -0.323093199
#> 50    10    5  0.955598763
#> 51     1    6  0.244890392
#> 52     2    6  2.372456761
#> 53     3    6  0.778353029
#> 54     4    6 -0.331724085
#> 55     5    6  0.616342832
#> 56     6    6  1.837246057
#> 57     7    6 -0.719890376
#> 58     8    6 -1.589074248
#> 59     9    6 -0.884380986
#> 60    10    6 -1.035043979
#> 61     1    7  1.127824366
#> 62     2    7  0.962778465
#> 63     3    7 -0.254824819
#> 64     4    7  0.811432116
#> 65     5    7  0.776670865
#> 66     6    7  0.132104872
#> 67     7    7 -2.889794544
#> 68     8    7  1.040517921
#> 69     9    7 -0.296165947
#> 70    10    7  2.550532779
#> 71     1    8  0.854239017
#> 72     2    8  0.649241597
#> 73     3    8  1.106744744
#> 74     4    8 -0.696491442
#> 75     5    8  0.265423068
#> 76     6    8 -2.570509708
#> 77     7    8 -1.621272905
#> 78     8    8 -0.270104703
#> 79     9    8  0.655654553
#> 80    10    8 -0.459743595
#> 81     1    9  2.644803828
#> 82     2    9 -0.004005489
#> 83     3    9  1.366888857
#> 84     4    9 -1.126319975
#> 85     5    9 -1.754619361
#> 86     6    9 -1.675729147
#> 87     7    9  0.234049761
#> 88     8    9 -0.926892603
#> 89     9    9 -2.714641135
#> 90    10    9 -3.185854788
#> 91     1   10  0.726258915
#> 92     2   10  2.577804207
#> 93     3   10  1.389090882
#> 94     4   10 -1.275093712
#> 95     5   10 -1.103043101
#> 96     6   10  0.973944225
#> 97     7   10  0.468602907
#> 98     8   10 -1.189214063
#> 99     9   10 -0.680780723
#> 100   10   10 -1.323711961