Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2            z
#> 1      1    1  3.083833867
#> 2      2    1  2.638842189
#> 3      3    1  3.373590656
#> 4      4    1  3.902118872
#> 5      5    1  1.953499684
#> 6      6    1  2.629387813
#> 7      7    1  4.852024565
#> 8      8    1  1.327024101
#> 9      9    1  2.326885381
#> 10    10    1  4.863623272
#> 11     1    2  3.999094820
#> 12     2    2  2.974126903
#> 13     3    2  5.201050199
#> 14     4    2  4.700422626
#> 15     5    2  3.440106887
#> 16     6    2  2.200791394
#> 17     7    2  3.384411992
#> 18     8    2  1.739321758
#> 19     9    2  3.090139573
#> 20    10    2  4.116451657
#> 21     1    3  1.712120971
#> 22     2    3  3.283688090
#> 23     3    3  4.832983956
#> 24     4    3  3.707891698
#> 25     5    3  3.194050097
#> 26     6    3  4.649961947
#> 27     7    3  2.076422307
#> 28     8    3  4.325790406
#> 29     9    3  2.953460962
#> 30    10    3  1.573523546
#> 31     1    4  1.676388952
#> 32     2    4  4.630847668
#> 33     3    4  6.042809276
#> 34     4    4  3.326580570
#> 35     5    4  3.373892688
#> 36     6    4  4.124827840
#> 37     7    4  2.544573854
#> 38     8    4  4.537951949
#> 39     9    4  1.174445500
#> 40    10    4  1.234838033
#> 41     1    5  3.695933915
#> 42     2    5  2.967763445
#> 43     3    5  1.902405401
#> 44     4    5  2.559499321
#> 45     5    5  3.470179858
#> 46     6    5  4.353619168
#> 47     7    5  4.249015598
#> 48     8    5  3.293623499
#> 49     9    5  0.004792937
#> 50    10    5  2.734111173
#> 51     1    6  3.078991185
#> 52     2    6  2.968434791
#> 53     3    6  3.441910079
#> 54     4    6  1.613819075
#> 55     5    6  3.427330056
#> 56     6    6  4.080886099
#> 57     7    6  4.077243874
#> 58     8    6  1.835525010
#> 59     9    6  1.367462320
#> 60    10    6  1.185378934
#> 61     1    7 -0.578276574
#> 62     2    7  1.766335392
#> 63     3    7  5.035916952
#> 64     4    7  2.897533504
#> 65     5    7  2.703468988
#> 66     6    7  4.078414434
#> 67     7    7  3.659455864
#> 68     8    7  2.392301719
#> 69     9    7  0.231954381
#> 70    10    7  0.420060930
#> 71     1    8  1.409971254
#> 72     2    8  1.186282806
#> 73     3    8  4.153895919
#> 74     4    8  3.477674991
#> 75     5    8  1.669329400
#> 76     6    8  2.273201032
#> 77     7    8  0.361050135
#> 78     8    8  2.581206938
#> 79     9    8  1.807476874
#> 80    10    8  3.868124016
#> 81     1    9  4.393654611
#> 82     2    9  2.473416209
#> 83     3    9  3.427415018
#> 84     4    9  1.914739202
#> 85     5    9  2.841550507
#> 86     6    9  1.788028496
#> 87     7    9  3.089651395
#> 88     8    9  1.095932915
#> 89     9    9  1.201527889
#> 90    10    9  1.754073292
#> 91     1   10  3.305349937
#> 92     2   10  3.167926379
#> 93     3   10  1.249296146
#> 94     4   10  0.463839702
#> 95     5   10  1.661361346
#> 96     6   10  2.305908648
#> 97     7   10  3.108598132
#> 98     8   10  1.579880293
#> 99     9   10  0.280259042
#> 100   10   10  1.578555011