Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -0.72289961
#> 2      2    1 -0.76902971
#> 3      3    1 -0.05573928
#> 4      4    1  0.88898620
#> 5      5    1 -1.05228623
#> 6      6    1 -2.89248752
#> 7      7    1 -2.21612259
#> 8      8    1 -1.40550655
#> 9      9    1 -2.71792624
#> 10    10    1 -4.25439437
#> 11     1    2  1.07608862
#> 12     2    2 -0.27642551
#> 13     3    2  0.07602185
#> 14     4    2 -0.27333835
#> 15     5    2  1.10157726
#> 16     6    2  1.04982695
#> 17     7    2 -2.76149071
#> 18     8    2 -1.04348623
#> 19     9    2 -2.74274965
#> 20    10    2 -0.66035073
#> 21     1    3  0.34979638
#> 22     2    3  0.24917179
#> 23     3    3 -0.24481153
#> 24     4    3 -0.23683567
#> 25     5    3  0.14739382
#> 26     6    3 -0.50611987
#> 27     7    3 -1.74679951
#> 28     8    3 -0.79809646
#> 29     9    3 -1.40842424
#> 30    10    3 -0.34462900
#> 31     1    4 -1.56818668
#> 32     2    4 -0.36354726
#> 33     3    4 -2.00643726
#> 34     4    4 -0.88580967
#> 35     5    4  0.97483685
#> 36     6    4 -1.04536262
#> 37     7    4 -1.25405025
#> 38     8    4 -2.28272195
#> 39     9    4 -1.07377458
#> 40    10    4 -0.83251583
#> 41     1    5  0.03371217
#> 42     2    5  1.33605205
#> 43     3    5  0.24601267
#> 44     4    5  0.69342215
#> 45     5    5  0.78776165
#> 46     6    5 -0.88206042
#> 47     7    5 -2.34835129
#> 48     8    5 -2.27155663
#> 49     9    5 -1.55198944
#> 50    10    5 -0.91419767
#> 51     1    6  1.76265565
#> 52     2    6  1.79749352
#> 53     3    6  1.38818825
#> 54     4    6  1.08974511
#> 55     5    6  0.29200914
#> 56     6    6 -1.31781960
#> 57     7    6 -1.95126774
#> 58     8    6 -0.04199442
#> 59     9    6 -2.16095064
#> 60    10    6  0.15885784
#> 61     1    7  0.96388942
#> 62     2    7  1.65416565
#> 63     3    7 -0.01246636
#> 64     4    7  1.25263197
#> 65     5    7  1.03323329
#> 66     6    7 -1.70222163
#> 67     7    7 -0.63365703
#> 68     8    7 -2.64415396
#> 69     9    7 -1.07153326
#> 70    10    7 -1.65324367
#> 71     1    8 -0.37985719
#> 72     2    8  0.74702518
#> 73     3    8 -0.75515773
#> 74     4    8 -1.15307245
#> 75     5    8  3.05552129
#> 76     6    8  0.35532197
#> 77     7    8 -0.55843697
#> 78     8    8 -0.25932149
#> 79     9    8 -0.46542269
#> 80    10    8 -2.78876490
#> 81     1    9  0.14879163
#> 82     2    9  1.22708417
#> 83     3    9 -0.23830056
#> 84     4    9 -0.66466424
#> 85     5    9 -1.17122242
#> 86     6    9  2.03556078
#> 87     7    9 -0.59702734
#> 88     8    9  0.19308632
#> 89     9    9 -0.04063284
#> 90    10    9  0.93508253
#> 91     1   10 -0.76557845
#> 92     2   10  0.55060019
#> 93     3   10 -0.42429557
#> 94     4   10  0.75612401
#> 95     5   10  0.33703158
#> 96     6   10  0.49065868
#> 97     7   10  0.45558121
#> 98     8   10  0.67928327
#> 99     9   10  1.28054445
#> 100   10   10 -0.28082513