Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -2.09384826
#> 2      2    1 -0.49100012
#> 3      3    1  1.22488074
#> 4      4    1 -0.44504106
#> 5      5    1 -0.03018957
#> 6      6    1 -1.04476662
#> 7      7    1 -1.26244106
#> 8      8    1  0.10274246
#> 9      9    1  0.30153495
#> 10    10    1 -0.12903286
#> 11     1    2  2.15768240
#> 12     2    2  1.09033399
#> 13     3    2 -2.44774293
#> 14     4    2 -1.10489094
#> 15     5    2  0.64878753
#> 16     6    2  1.18758739
#> 17     7    2  0.33392364
#> 18     8    2 -1.20187269
#> 19     9    2 -2.96300428
#> 20    10    2  0.02295197
#> 21     1    3 -0.03239508
#> 22     2    3 -1.32886871
#> 23     3    3 -1.77288838
#> 24     4    3 -2.14193372
#> 25     5    3 -3.17103076
#> 26     6    3 -1.71953332
#> 27     7    3  0.08617412
#> 28     8    3 -1.08549984
#> 29     9    3 -3.14328615
#> 30    10    3 -2.69110912
#> 31     1    4  1.39925593
#> 32     2    4  0.19862287
#> 33     3    4 -3.25704263
#> 34     4    4 -2.28001026
#> 35     5    4 -2.75294581
#> 36     6    4 -1.03955377
#> 37     7    4 -1.50110889
#> 38     8    4 -1.80655532
#> 39     9    4 -1.71374549
#> 40    10    4 -1.86984288
#> 41     1    5  0.49013650
#> 42     2    5  0.83145948
#> 43     3    5  0.15110686
#> 44     4    5 -3.11217256
#> 45     5    5 -5.38200799
#> 46     6    5 -3.54583471
#> 47     7    5 -0.14661266
#> 48     8    5 -0.26714934
#> 49     9    5  0.99476419
#> 50    10    5  0.58128560
#> 51     1    6  0.60707164
#> 52     2    6 -1.38715836
#> 53     3    6 -2.30054810
#> 54     4    6 -1.59126910
#> 55     5    6 -0.46683498
#> 56     6    6 -2.96043294
#> 57     7    6 -1.46629363
#> 58     8    6 -0.68631390
#> 59     9    6 -0.88798089
#> 60    10    6 -1.79532137
#> 61     1    7 -0.46265807
#> 62     2    7 -0.68083161
#> 63     3    7 -0.23034242
#> 64     4    7 -1.53932388
#> 65     5    7 -2.21200473
#> 66     6    7 -2.23211386
#> 67     7    7 -1.98852848
#> 68     8    7 -2.68654008
#> 69     9    7 -1.95886181
#> 70    10    7 -1.39769435
#> 71     1    8  0.33193915
#> 72     2    8 -0.33036263
#> 73     3    8  0.01039195
#> 74     4    8 -2.08688884
#> 75     5    8 -2.75450571
#> 76     6    8 -2.48713633
#> 77     7    8 -0.81311333
#> 78     8    8 -1.20941868
#> 79     9    8 -0.66619824
#> 80    10    8 -1.29677279
#> 81     1    9 -0.40223191
#> 82     2    9 -1.36953993
#> 83     3    9 -2.27507432
#> 84     4    9 -0.87112128
#> 85     5    9 -1.70208661
#> 86     6    9 -3.07378861
#> 87     7    9 -2.01941785
#> 88     8    9 -0.77339067
#> 89     9    9 -0.72331865
#> 90    10    9  0.60330421
#> 91     1   10 -2.13400296
#> 92     2   10  0.06324827
#> 93     3   10  0.83996280
#> 94     4   10 -1.13486401
#> 95     5   10 -2.20542794
#> 96     6   10  0.66587766
#> 97     7   10  0.06524872
#> 98     8   10 -2.30567134
#> 99     9   10 -0.53281295
#> 100   10   10 -1.06417641