Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1  1.56955480
#> 2      2    1  1.22581076
#> 3      3    1  0.50729897
#> 4      4    1  1.84169994
#> 5      5    1  0.46293680
#> 6      6    1  1.38778831
#> 7      7    1  0.16798327
#> 8      8    1  2.80499291
#> 9      9    1 -1.83534669
#> 10    10    1 -0.20857499
#> 11     1    2 -1.67932530
#> 12     2    2 -0.54305620
#> 13     3    2  0.85912393
#> 14     4    2  0.64851670
#> 15     5    2  1.95019086
#> 16     6    2 -0.71780787
#> 17     7    2  0.28760221
#> 18     8    2  0.32291879
#> 19     9    2  0.11539943
#> 20    10    2 -0.43971280
#> 21     1    3  2.15134305
#> 22     2    3 -1.13802823
#> 23     3    3  0.49286027
#> 24     4    3  2.46373109
#> 25     5    3  0.46460824
#> 26     6    3  3.27383621
#> 27     7    3  2.05743884
#> 28     8    3  2.40163681
#> 29     9    3  0.84650677
#> 30    10    3  1.99373665
#> 31     1    4 -0.61953631
#> 32     2    4  0.71643797
#> 33     3    4  4.10507873
#> 34     4    4  1.44257334
#> 35     5    4  1.27273075
#> 36     6    4  0.93971834
#> 37     7    4  1.06148376
#> 38     8    4  1.30124777
#> 39     9    4  1.48006617
#> 40    10    4  1.16779848
#> 41     1    5 -2.36321120
#> 42     2    5  0.16371324
#> 43     3    5  3.07977614
#> 44     4    5  2.79522318
#> 45     5    5  1.74991046
#> 46     6    5  0.94456285
#> 47     7    5  2.15824977
#> 48     8    5  0.23350304
#> 49     9    5  0.19140367
#> 50    10    5  2.40826002
#> 51     1    6  0.93604739
#> 52     2    6  3.21692974
#> 53     3    6  5.05798425
#> 54     4    6  1.01676608
#> 55     5    6  0.04311057
#> 56     6    6  1.30518990
#> 57     7    6  2.40064212
#> 58     8    6  2.70606686
#> 59     9    6  1.67929205
#> 60    10    6  1.40920879
#> 61     1    7  2.64232753
#> 62     2    7  2.71522018
#> 63     3    7  3.25240443
#> 64     4    7  2.46352970
#> 65     5    7  3.56469700
#> 66     6    7  3.40639707
#> 67     7    7  3.44004822
#> 68     8    7  1.05895758
#> 69     9    7  4.62454820
#> 70    10    7  1.75057744
#> 71     1    8  2.26441291
#> 72     2    8  3.91238285
#> 73     3    8  2.82993652
#> 74     4    8  3.72586311
#> 75     5    8  1.15627881
#> 76     6    8  0.67682797
#> 77     7    8  2.01923700
#> 78     8    8  2.15097867
#> 79     9    8  1.86142910
#> 80    10    8 -0.26253641
#> 81     1    9  2.79004686
#> 82     2    9  2.31334390
#> 83     3    9  1.65542775
#> 84     4    9  3.59331160
#> 85     5    9  1.16559420
#> 86     6    9  5.35764526
#> 87     7    9  3.10383299
#> 88     8    9  2.31854103
#> 89     9    9  1.78852773
#> 90    10    9  2.38058100
#> 91     1   10  2.11217706
#> 92     2   10  3.12820727
#> 93     3   10  5.12901648
#> 94     4   10  4.02852113
#> 95     5   10  4.06559586
#> 96     6   10  3.66922503
#> 97     7   10  1.57725270
#> 98     8   10  2.03906643
#> 99     9   10  0.97210196
#> 100   10   10  2.63125143