Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1  4.44654334
#> 2      2    1  1.26156185
#> 3      3    1  0.83693957
#> 4      4    1  3.65364875
#> 5      5    1  3.43354165
#> 6      6    1  1.72160697
#> 7      7    1  0.92734566
#> 8      8    1 -1.37881024
#> 9      9    1  0.03285678
#> 10    10    1 -0.23684981
#> 11     1    2  3.36278826
#> 12     2    2  2.47618750
#> 13     3    2  3.47063711
#> 14     4    2 -0.23409592
#> 15     5    2  0.21592817
#> 16     6    2  2.76995371
#> 17     7    2  0.22118710
#> 18     8    2  3.08670230
#> 19     9    2 -1.35769816
#> 20    10    2  0.02583270
#> 21     1    3  2.52541092
#> 22     2    3 -0.33121849
#> 23     3    3  3.29316389
#> 24     4    3  0.45591104
#> 25     5    3  2.13755987
#> 26     6    3  1.32019092
#> 27     7    3  2.77095806
#> 28     8    3 -0.17339589
#> 29     9    3  0.50527225
#> 30    10    3  1.22311117
#> 31     1    4  1.65547175
#> 32     2    4  3.12689477
#> 33     3    4  0.68987566
#> 34     4    4  0.64390962
#> 35     5    4  2.92530284
#> 36     6    4  0.40101964
#> 37     7    4 -1.23838851
#> 38     8    4 -2.12099037
#> 39     9    4 -0.30393012
#> 40    10    4 -1.44597184
#> 41     1    5 -1.16649427
#> 42     2    5 -0.39860059
#> 43     3    5 -1.65843558
#> 44     4    5  1.09326945
#> 45     5    5  2.39122315
#> 46     6    5 -0.60534193
#> 47     7    5  0.01447877
#> 48     8    5 -0.27845715
#> 49     9    5 -0.98077894
#> 50    10    5  0.95652412
#> 51     1    6  0.23261273
#> 52     2    6  1.06079716
#> 53     3    6 -0.72199995
#> 54     4    6  2.10037524
#> 55     5    6  0.67697914
#> 56     6    6  0.69755093
#> 57     7    6  0.99666389
#> 58     8    6 -1.57030330
#> 59     9    6 -2.92789882
#> 60    10    6 -2.36845138
#> 61     1    7  1.00205861
#> 62     2    7  0.39051126
#> 63     3    7 -0.34939426
#> 64     4    7  0.89580371
#> 65     5    7 -0.15533365
#> 66     6    7 -0.92262758
#> 67     7    7  0.28013532
#> 68     8    7 -0.80111277
#> 69     9    7 -2.00730797
#> 70    10    7 -0.66284194
#> 71     1    8  1.58767573
#> 72     2    8  1.91863092
#> 73     3    8  2.28250008
#> 74     4    8  1.37925656
#> 75     5    8 -1.89009595
#> 76     6    8  0.35421592
#> 77     7    8 -0.76011954
#> 78     8    8 -0.91073504
#> 79     9    8 -1.10381024
#> 80    10    8 -2.82569631
#> 81     1    9  0.34593480
#> 82     2    9  0.56843923
#> 83     3    9  1.11636579
#> 84     4    9 -2.60597539
#> 85     5    9 -2.80003996
#> 86     6    9 -1.82596211
#> 87     7    9 -2.46342631
#> 88     8    9  0.04367252
#> 89     9    9 -1.82241749
#> 90    10    9 -3.76186027
#> 91     1   10  2.18698829
#> 92     2   10 -0.69917643
#> 93     3   10 -0.38195190
#> 94     4   10 -2.19660169
#> 95     5   10 -1.23533260
#> 96     6   10 -0.59325576
#> 97     7   10  0.49148096
#> 98     8   10 -3.23643294
#> 99     9   10 -3.29368315
#> 100   10   10 -0.45090525