Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -0.40795135
#> 2      2    1  2.09640619
#> 3      3    1 -1.96306603
#> 4      4    1  2.71240684
#> 5      5    1  1.77620176
#> 6      6    1  1.51618641
#> 7      7    1  2.86130979
#> 8      8    1  2.77935147
#> 9      9    1  1.79759388
#> 10    10    1  1.80245378
#> 11     1    2 -1.09435760
#> 12     2    2  2.57475960
#> 13     3    2 -2.28271330
#> 14     4    2  1.54925623
#> 15     5    2 -0.45532523
#> 16     6    2  1.12521758
#> 17     7    2  1.67431507
#> 18     8    2  1.36890600
#> 19     9    2  1.46818822
#> 20    10    2  2.26839298
#> 21     1    3  0.20504038
#> 22     2    3  2.42351234
#> 23     3    3 -0.29673816
#> 24     4    3  0.47579379
#> 25     5    3  1.56950903
#> 26     6    3  2.88068806
#> 27     7    3  1.99308650
#> 28     8    3  2.07792629
#> 29     9    3  2.32312077
#> 30    10    3 -0.52106757
#> 31     1    4  1.27616291
#> 32     2    4  1.36809407
#> 33     3    4  0.95937551
#> 34     4    4  1.06420904
#> 35     5    4  2.68384392
#> 36     6    4  0.90207174
#> 37     7    4  1.84283296
#> 38     8    4  0.52212896
#> 39     9    4 -1.82339416
#> 40    10    4  0.28330084
#> 41     1    5  1.07481628
#> 42     2    5  0.62466989
#> 43     3    5  1.43720349
#> 44     4    5 -0.86041267
#> 45     5    5  0.88606772
#> 46     6    5  3.65074019
#> 47     7    5  3.50979802
#> 48     8    5 -0.03678949
#> 49     9    5  1.55738332
#> 50    10    5 -0.36906145
#> 51     1    6 -0.01153720
#> 52     2    6 -0.38035648
#> 53     3    6  1.88813009
#> 54     4    6  0.82382387
#> 55     5    6  2.98435230
#> 56     6    6  1.52709338
#> 57     7    6  1.55409128
#> 58     8    6  0.24767535
#> 59     9    6  0.89970189
#> 60    10    6  0.32449890
#> 61     1    7  0.02553897
#> 62     2    7 -0.18971765
#> 63     3    7  1.83545750
#> 64     4    7 -0.35884771
#> 65     5    7  2.40779665
#> 66     6    7  0.37161663
#> 67     7    7  1.93294171
#> 68     8    7  0.03450143
#> 69     9    7 -0.68190548
#> 70    10    7 -0.04517260
#> 71     1    8  1.09382577
#> 72     2    8  0.28478370
#> 73     3    8  0.02092053
#> 74     4    8  0.80576629
#> 75     5    8  0.82871694
#> 76     6    8  2.53858786
#> 77     7    8 -1.29994575
#> 78     8    8 -1.58810839
#> 79     9    8 -1.03753591
#> 80    10    8 -0.36664054
#> 81     1    9  1.24764862
#> 82     2    9 -0.01465044
#> 83     3    9 -1.03509941
#> 84     4    9  0.03690352
#> 85     5    9  1.44196933
#> 86     6    9  0.19834648
#> 87     7    9 -1.16768288
#> 88     8    9 -0.92776130
#> 89     9    9 -0.01640387
#> 90    10    9  0.75403861
#> 91     1   10  2.24471592
#> 92     2   10  1.50703010
#> 93     3   10 -0.19376450
#> 94     4   10 -2.01144818
#> 95     5   10  0.74730684
#> 96     6   10  0.75356851
#> 97     7   10  0.49788328
#> 98     8   10 -1.10415007
#> 99     9   10  0.27032492
#> 100   10   10 -1.12265239