Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -1.05909950
#> 2      2    1 -3.79453615
#> 3      3    1 -3.34336175
#> 4      4    1 -1.26644498
#> 5      5    1 -0.31124295
#> 6      6    1  1.19836974
#> 7      7    1 -0.17079544
#> 8      8    1  0.17410445
#> 9      9    1 -1.73510897
#> 10    10    1 -0.66247113
#> 11     1    2  1.28923863
#> 12     2    2 -1.88284786
#> 13     3    2 -0.33192526
#> 14     4    2 -1.42312213
#> 15     5    2 -1.56855788
#> 16     6    2  0.73287223
#> 17     7    2  0.19264988
#> 18     8    2  2.26594922
#> 19     9    2 -0.07048469
#> 20    10    2 -0.22164922
#> 21     1    3  0.32165942
#> 22     2    3 -1.95506364
#> 23     3    3 -0.91263560
#> 24     4    3 -2.42496230
#> 25     5    3 -0.28402293
#> 26     6    3  0.07383017
#> 27     7    3 -0.94704320
#> 28     8    3 -0.80431658
#> 29     9    3 -0.25366281
#> 30    10    3 -0.83717827
#> 31     1    4  0.08443432
#> 32     2    4  0.84799223
#> 33     3    4 -0.76327517
#> 34     4    4 -1.52565766
#> 35     5    4 -0.40133575
#> 36     6    4  1.60431060
#> 37     7    4  0.66331226
#> 38     8    4  0.11611103
#> 39     9    4 -2.01184438
#> 40    10    4 -1.19333481
#> 41     1    5  2.09230793
#> 42     2    5  0.84218521
#> 43     3    5  0.14539917
#> 44     4    5 -1.41861411
#> 45     5    5 -0.49876149
#> 46     6    5 -0.51414809
#> 47     7    5 -1.21103151
#> 48     8    5 -0.66135653
#> 49     9    5 -0.13105581
#> 50    10    5  0.79887041
#> 51     1    6 -0.75400062
#> 52     2    6  2.19716078
#> 53     3    6  1.08135543
#> 54     4    6 -0.32961303
#> 55     5    6 -1.21906019
#> 56     6    6  0.91273732
#> 57     7    6  1.14140419
#> 58     8    6 -0.80971422
#> 59     9    6 -2.36727757
#> 60    10    6 -1.54362125
#> 61     1    7 -1.19391599
#> 62     2    7  0.70158520
#> 63     3    7  0.63745204
#> 64     4    7  0.81659617
#> 65     5    7 -0.67458525
#> 66     6    7 -1.47527159
#> 67     7    7 -1.25041939
#> 68     8    7 -0.83265454
#> 69     9    7 -0.11406920
#> 70    10    7 -0.30988793
#> 71     1    8 -2.95727285
#> 72     2    8 -1.66734463
#> 73     3    8  0.88641833
#> 74     4    8 -1.61653653
#> 75     5    8 -0.93504887
#> 76     6    8 -1.57223678
#> 77     7    8 -3.97807506
#> 78     8    8 -2.56703543
#> 79     9    8 -0.85249097
#> 80    10    8 -1.17658276
#> 81     1    9 -2.46111561
#> 82     2    9 -2.21516468
#> 83     3    9 -0.98019549
#> 84     4    9 -0.84292079
#> 85     5    9 -2.39472699
#> 86     6    9 -1.55045559
#> 87     7    9 -1.69727078
#> 88     8    9 -0.95813180
#> 89     9    9 -0.16456063
#> 90    10    9 -1.56152836
#> 91     1   10 -0.29951974
#> 92     2   10 -0.27558641
#> 93     3   10 -3.11002900
#> 94     4   10 -2.34583130
#> 95     5   10 -1.71747327
#> 96     6   10 -1.65897814
#> 97     7   10 -0.36177522
#> 98     8   10 -3.06400908
#> 99     9   10 -2.58960167
#> 100   10   10 -2.78532321