Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -0.93385473
#> 2      2    1 -0.56263870
#> 3      3    1  1.46884684
#> 4      4    1  2.10394748
#> 5      5    1  0.82814321
#> 6      6    1  1.41863766
#> 7      7    1  2.65195790
#> 8      8    1  3.60582536
#> 9      9    1  0.61048325
#> 10    10    1  0.66506311
#> 11     1    2 -0.08001172
#> 12     2    2  1.89611014
#> 13     3    2  2.85943041
#> 14     4    2  2.80149618
#> 15     5    2  3.19857655
#> 16     6    2  3.03387629
#> 17     7    2  2.70414107
#> 18     8    2  1.92725748
#> 19     9    2 -0.57670952
#> 20    10    2  1.62049959
#> 21     1    3  2.57336598
#> 22     2    3  3.67786332
#> 23     3    3  2.51360304
#> 24     4    3  1.98831849
#> 25     5    3  0.75881076
#> 26     6    3  2.87779445
#> 27     7    3  3.59844503
#> 28     8    3  3.08686153
#> 29     9    3  2.36825636
#> 30    10    3  2.62860633
#> 31     1    4  2.06981321
#> 32     2    4  3.15647653
#> 33     3    4  4.43156168
#> 34     4    4  3.59615547
#> 35     5    4  2.55919157
#> 36     6    4  0.81757979
#> 37     7    4  1.84113186
#> 38     8    4  0.65428255
#> 39     9    4  2.64327221
#> 40    10    4  0.97706746
#> 41     1    5  1.89261144
#> 42     2    5  1.04155976
#> 43     3    5  1.77248137
#> 44     4    5  2.09647133
#> 45     5    5  2.65776991
#> 46     6    5 -0.05623807
#> 47     7    5  4.13736357
#> 48     8    5  4.55020438
#> 49     9    5  0.70058537
#> 50    10    5 -0.18850517
#> 51     1    6  0.57549911
#> 52     2    6  3.03696366
#> 53     3    6  3.87365168
#> 54     4    6  6.20913206
#> 55     5    6  2.49434652
#> 56     6    6  1.30448017
#> 57     7    6  2.20212224
#> 58     8    6  0.61724213
#> 59     9    6  3.59867183
#> 60    10    6  1.65493666
#> 61     1    7  4.62988520
#> 62     2    7  2.80783649
#> 63     3    7  3.59855229
#> 64     4    7  3.07996792
#> 65     5    7  2.98185746
#> 66     6    7  4.52917222
#> 67     7    7  1.64551655
#> 68     8    7  3.40258024
#> 69     9    7  2.28011393
#> 70    10    7  1.05704036
#> 71     1    8  2.03908316
#> 72     2    8  2.34151150
#> 73     3    8  1.88664381
#> 74     4    8  2.48977968
#> 75     5    8  3.97841955
#> 76     6    8  2.93008210
#> 77     7    8  2.08350832
#> 78     8    8  3.99659073
#> 79     9    8  6.49733672
#> 80    10    8  3.25914581
#> 81     1    9  3.30244445
#> 82     2    9  1.03118661
#> 83     3    9  2.62667752
#> 84     4    9  3.52109656
#> 85     5    9  5.37993104
#> 86     6    9  5.40376954
#> 87     7    9  3.01994034
#> 88     8    9  5.38730431
#> 89     9    9  3.79976981
#> 90    10    9  3.15635637
#> 91     1   10  3.21098286
#> 92     2   10  1.43280703
#> 93     3   10  2.36132628
#> 94     4   10  2.67633563
#> 95     5   10  2.00645107
#> 96     6   10  4.48156495
#> 97     7   10  2.77836619
#> 98     8   10  2.95576115
#> 99     9   10  3.19921041
#> 100   10   10  4.76699448