Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2             z
#> 1      1    1  0.0684688933
#> 2      2    1  0.5363340447
#> 3      3    1  1.3882697439
#> 4      4    1  0.5209691655
#> 5      5    1  0.9240000260
#> 6      6    1  1.6824417414
#> 7      7    1  2.3013970951
#> 8      8    1  0.2763336570
#> 9      9    1  1.6931155189
#> 10    10    1  2.1507445186
#> 11     1    2  0.0194743901
#> 12     2    2  1.4224417241
#> 13     3    2  0.8795238446
#> 14     4    2  2.6727770946
#> 15     5    2  1.2723673562
#> 16     6    2  1.4160487633
#> 17     7    2  0.8186376538
#> 18     8    2  1.6035848026
#> 19     9    2  2.1476248091
#> 20    10    2  3.0915362157
#> 21     1    3 -0.2256085489
#> 22     2    3  3.4509075581
#> 23     3    3  1.5877619063
#> 24     4    3  0.6892807899
#> 25     5    3  3.6776036042
#> 26     6    3  2.1587449377
#> 27     7    3  1.5507813566
#> 28     8    3  3.5526962442
#> 29     9    3  3.3168718426
#> 30    10    3  3.9316645365
#> 31     1    4  2.1168322292
#> 32     2    4  0.7430066548
#> 33     3    4  1.5003962948
#> 34     4    4  1.7473771849
#> 35     5    4  3.2005306912
#> 36     6    4  4.2944142095
#> 37     7    4  4.9399789018
#> 38     8    4  1.9070334051
#> 39     9    4  3.6310088930
#> 40    10    4  2.3512890100
#> 41     1    5  0.0238019415
#> 42     2    5  0.5225509200
#> 43     3    5  2.6870944911
#> 44     4    5  1.0355118527
#> 45     5    5  1.7263265684
#> 46     6    5  2.8073103635
#> 47     7    5  2.8411006696
#> 48     8    5  2.8477805366
#> 49     9    5  5.2711443283
#> 50    10    5  2.5806414019
#> 51     1    6  1.8289684086
#> 52     2    6 -0.3492504817
#> 53     3    6  3.2284674184
#> 54     4    6  1.7913757957
#> 55     5    6  3.9688088252
#> 56     6    6  3.6474392809
#> 57     7    6  2.4768657297
#> 58     8    6  4.0543615849
#> 59     9    6  2.9926004258
#> 60    10    6  4.3810944168
#> 61     1    7 -0.1601036502
#> 62     2    7 -0.0006038604
#> 63     3    7  1.7742493656
#> 64     4    7  3.4761382939
#> 65     5    7  0.2359027228
#> 66     6    7  2.2939079576
#> 67     7    7  4.0530553249
#> 68     8    7  3.1257635445
#> 69     9    7  3.5082102895
#> 70    10    7  3.8569905625
#> 71     1    8  1.4988406649
#> 72     2    8  1.8470725493
#> 73     3    8  0.8226487179
#> 74     4    8  4.4758987481
#> 75     5    8  1.8382704017
#> 76     6    8  2.8728642451
#> 77     7    8  2.9727350995
#> 78     8    8  2.2291227943
#> 79     9    8  2.7958548356
#> 80    10    8  1.7224636304
#> 81     1    9  1.5911593413
#> 82     2    9  2.9614401665
#> 83     3    9 -0.2283124061
#> 84     4    9  3.2765303069
#> 85     5    9  2.3517438700
#> 86     6    9  0.9383833197
#> 87     7    9  3.2768701222
#> 88     8    9  3.1738983525
#> 89     9    9  2.7565692761
#> 90    10    9  2.4136650704
#> 91     1   10 -0.5115157481
#> 92     2   10  0.8930639955
#> 93     3   10  2.6801951906
#> 94     4   10  2.6653227494
#> 95     5   10  4.5117049283
#> 96     6   10  3.4276197990
#> 97     7   10  2.9257780892
#> 98     8   10  3.1592843516
#> 99     9   10  2.7597100865
#> 100   10   10  2.2537760575