Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1 -3.14728929
#> 2      2    1 -2.99513973
#> 3      3    1 -0.77600729
#> 4      4    1 -2.39528166
#> 5      5    1 -1.50615662
#> 6      6    1 -0.62637789
#> 7      7    1 -2.34425815
#> 8      8    1  0.43288504
#> 9      9    1 -0.74124727
#> 10    10    1 -0.78171054
#> 11     1    2 -1.58037451
#> 12     2    2 -2.41447686
#> 13     3    2 -1.89363870
#> 14     4    2 -0.93089223
#> 15     5    2 -2.19858846
#> 16     6    2 -3.43999910
#> 17     7    2 -1.89139091
#> 18     8    2 -0.96005585
#> 19     9    2 -0.44091010
#> 20    10    2  1.08488818
#> 21     1    3 -0.94354065
#> 22     2    3 -1.30541509
#> 23     3    3 -0.36405492
#> 24     4    3 -0.84804009
#> 25     5    3 -1.95119725
#> 26     6    3 -3.38488113
#> 27     7    3 -2.37127487
#> 28     8    3 -2.34464405
#> 29     9    3  0.15866100
#> 30    10    3 -1.33874233
#> 31     1    4 -1.81066703
#> 32     2    4 -1.93648743
#> 33     3    4  0.15447006
#> 34     4    4  1.91610193
#> 35     5    4 -2.11461230
#> 36     6    4 -1.47804659
#> 37     7    4 -3.30745684
#> 38     8    4 -0.85225492
#> 39     9    4 -1.78026312
#> 40    10    4 -2.68977600
#> 41     1    5 -1.01673650
#> 42     2    5 -0.35192342
#> 43     3    5 -0.03720686
#> 44     4    5  0.14808582
#> 45     5    5  1.07331169
#> 46     6    5 -0.57689174
#> 47     7    5 -1.66655331
#> 48     8    5  0.66250560
#> 49     9    5 -0.70310469
#> 50    10    5 -0.56117016
#> 51     1    6 -2.90465057
#> 52     2    6 -0.90161545
#> 53     3    6 -0.27901638
#> 54     4    6  0.87774073
#> 55     5    6  0.96981541
#> 56     6    6  1.22710141
#> 57     7    6 -0.84977861
#> 58     8    6 -1.89236316
#> 59     9    6 -0.73652868
#> 60    10    6 -2.33316826
#> 61     1    7 -1.00909730
#> 62     2    7 -1.36680233
#> 63     3    7  0.55225770
#> 64     4    7  0.03695928
#> 65     5    7 -0.51217025
#> 66     6    7 -0.79596580
#> 67     7    7  0.16602849
#> 68     8    7 -2.31570340
#> 69     9    7 -1.47531231
#> 70    10    7 -0.63562165
#> 71     1    8 -0.24071467
#> 72     2    8  1.08600671
#> 73     3    8 -0.15162956
#> 74     4    8  0.09272995
#> 75     5    8 -0.60226190
#> 76     6    8 -0.15720701
#> 77     7    8 -2.54080462
#> 78     8    8 -0.62859159
#> 79     9    8 -1.90373426
#> 80    10    8 -2.04450807
#> 81     1    9  1.34145561
#> 82     2    9  0.71748029
#> 83     3    9 -0.66661355
#> 84     4    9  0.75078378
#> 85     5    9 -0.27146183
#> 86     6    9 -5.17200225
#> 87     7    9 -2.46053967
#> 88     8    9 -1.22715152
#> 89     9    9 -1.24830090
#> 90    10    9 -1.66185928
#> 91     1   10 -0.70773906
#> 92     2   10 -0.90211698
#> 93     3   10  0.18965358
#> 94     4   10  1.33876671
#> 95     5   10 -2.06641529
#> 96     6   10 -0.41059977
#> 97     7   10 -1.09434709
#> 98     8   10 -1.15490981
#> 99     9   10 -2.46713614
#> 100   10   10 -0.29447647