Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data
data.frame with x- and y-coordinates of locations for simulated data

xcol
name of the column in loc.data with x-coordinates, default is "x"

ycol
name of the column loc.data with y-coordinates, default is "y"

parsil
partial sill of autocorrelation model, default = 1

range
range of autocorrelation model, default = 1

nugget
range of autocorrelation model, default = 0

minorp
proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate
rotation of anisotropic axes, default = 90

extrap
extra covariance paramter

CorModel
autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")

#>     Var1 Var2           z
#> 1      1    1 -3.13059247
#> 2      2    1 -2.01194179
#> 3      3    1 -5.40122622
#> 4      4    1 -3.44330189
#> 5      5    1 -2.44165102
#> 6      6    1 -1.23423343
#> 7      7    1 -4.42156981
#> 8      8    1 -3.23535786
#> 9      9    1 -3.22648109
#> 10    10    1 -3.27157616
#> 11     1    2 -3.62170885
#> 12     2    2 -2.48556934
#> 13     3    2 -0.73797421
#> 14     4    2 -4.59561184
#> 15     5    2 -2.41420680
#> 16     6    2 -5.04270858
#> 17     7    2 -4.02222160
#> 18     8    2 -3.51092730
#> 19     9    2 -2.66685165
#> 20    10    2 -4.15020447
#> 21     1    3 -2.23376348
#> 22     2    3 -2.24736404
#> 23     3    3 -4.26233630
#> 24     4    3 -2.40953196
#> 25     5    3 -0.99145780
#> 26     6    3 -3.11933738
#> 27     7    3 -4.33908003
#> 28     8    3 -3.47790703
#> 29     9    3 -4.35001174
#> 30    10    3 -5.36919333
#> 31     1    4 -1.37581997
#> 32     2    4 -2.17105208
#> 33     3    4 -2.27196946
#> 34     4    4 -0.68852236
#> 35     5    4 -2.27288313
#> 36     6    4 -3.00215797
#> 37     7    4 -5.55168226
#> 38     8    4 -4.29841743
#> 39     9    4 -4.46495313
#> 40    10    4 -2.87816355
#> 41     1    5 -1.86858227
#> 42     2    5 -2.78160712
#> 43     3    5 -2.30994052
#> 44     4    5 -2.62386460
#> 45     5    5 -2.16225674
#> 46     6    5  0.17321596
#> 47     7    5 -2.92919402
#> 48     8    5 -2.71365443
#> 49     9    5 -3.25391958
#> 50    10    5 -5.67405106
#> 51     1    6 -0.89271870
#> 52     2    6 -2.10284524
#> 53     3    6 -2.23252952
#> 54     4    6 -2.03990122
#> 55     5    6 -2.03850007
#> 56     6    6 -1.57900369
#> 57     7    6 -5.20132081
#> 58     8    6 -0.29809702
#> 59     9    6 -3.51849202
#> 60    10    6 -3.27330054
#> 61     1    7 -0.27205758
#> 62     2    7 -2.25426778
#> 63     3    7 -0.53657597
#> 64     4    7 -1.98798573
#> 65     5    7 -3.31303997
#> 66     6    7 -3.54894654
#> 67     7    7 -1.53901955
#> 68     8    7 -2.42617475
#> 69     9    7 -2.23355724
#> 70    10    7 -4.99612287
#> 71     1    8 -3.09060018
#> 72     2    8 -1.51759365
#> 73     3    8 -0.25301913
#> 74     4    8 -1.48334355
#> 75     5    8 -0.51786911
#> 76     6    8 -1.39962484
#> 77     7    8 -1.38657558
#> 78     8    8 -2.36862130
#> 79     9    8 -2.46373135
#> 80    10    8 -3.12399689
#> 81     1    9  0.91193506
#> 82     2    9  0.70664743
#> 83     3    9  0.23316323
#> 84     4    9 -0.10010698
#> 85     5    9 -1.58445461
#> 86     6    9 -0.07852452
#> 87     7    9 -4.50338534
#> 88     8    9 -2.57149364
#> 89     9    9 -2.10528670
#> 90    10    9 -3.19091493
#> 91     1   10  0.36558483
#> 92     2   10  1.23344856
#> 93     3   10  1.25888298
#> 94     4   10 -0.20145389
#> 95     5   10 -0.12891716
#> 96     6   10  0.84362412
#> 97     7   10 -1.78006823
#> 98     8   10 -0.79736420
#> 99     9   10 -0.41634257
#> 100   10   10 -2.31871916