Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2            z
#> 1      1    1  1.038416490
#> 2      2    1 -0.471126320
#> 3      3    1 -1.667411598
#> 4      4    1  1.298771460
#> 5      5    1 -0.487296926
#> 6      6    1 -1.151277657
#> 7      7    1 -2.298625236
#> 8      8    1 -0.810536240
#> 9      9    1 -2.122589828
#> 10    10    1 -0.935802561
#> 11     1    2 -0.264263295
#> 12     2    2 -0.727034027
#> 13     3    2  1.464676279
#> 14     4    2 -2.403357123
#> 15     5    2 -0.949552051
#> 16     6    2 -0.640875398
#> 17     7    2 -0.006393008
#> 18     8    2 -1.388363210
#> 19     9    2  0.083576070
#> 20    10    2  0.438680733
#> 21     1    3 -0.263591707
#> 22     2    3  0.362093921
#> 23     3    3  0.106734212
#> 24     4    3  0.223098464
#> 25     5    3  1.348833327
#> 26     6    3  2.587807223
#> 27     7    3  0.932680310
#> 28     8    3  1.746438270
#> 29     9    3  2.428211593
#> 30    10    3  0.211567759
#> 31     1    4 -0.285708037
#> 32     2    4  1.185017778
#> 33     3    4 -1.556338221
#> 34     4    4 -0.158619836
#> 35     5    4 -0.399068580
#> 36     6    4  1.128818106
#> 37     7    4  0.826208595
#> 38     8    4  1.294247463
#> 39     9    4  1.069944320
#> 40    10    4  0.445630755
#> 41     1    5 -0.552311699
#> 42     2    5 -0.289806817
#> 43     3    5  0.665833261
#> 44     4    5  0.350314836
#> 45     5    5 -0.147680919
#> 46     6    5  0.800487080
#> 47     7    5  1.532777339
#> 48     8    5  0.944422484
#> 49     9    5  1.959658537
#> 50    10    5  2.434469912
#> 51     1    6 -0.292542365
#> 52     2    6  1.407796019
#> 53     3    6 -0.001165708
#> 54     4    6  1.932478372
#> 55     5    6  3.169824968
#> 56     6    6  0.973311642
#> 57     7    6  2.216533990
#> 58     8    6  1.903271868
#> 59     9    6  1.102973662
#> 60    10    6  1.385180645
#> 61     1    7 -1.135730413
#> 62     2    7  1.445645282
#> 63     3    7  0.599536253
#> 64     4    7  0.501302606
#> 65     5    7  2.712900258
#> 66     6    7  0.335394279
#> 67     7    7 -0.054698583
#> 68     8    7 -1.351833329
#> 69     9    7  0.033421920
#> 70    10    7  0.145623867
#> 71     1    8  1.679741552
#> 72     2    8 -0.800263860
#> 73     3    8  0.508065190
#> 74     4    8 -2.386498097
#> 75     5    8  0.569045884
#> 76     6    8  0.618277663
#> 77     7    8 -1.119653993
#> 78     8    8  2.073465170
#> 79     9    8  1.464899489
#> 80    10    8 -2.471278706
#> 81     1    9  1.865021104
#> 82     2    9  1.417836170
#> 83     3    9  1.373895180
#> 84     4    9  0.132062261
#> 85     5    9 -1.043122340
#> 86     6    9 -0.838027982
#> 87     7    9 -0.530508251
#> 88     8    9  1.424692371
#> 89     9    9  0.161558613
#> 90    10    9 -1.593798260
#> 91     1   10  1.286820581
#> 92     2   10  0.400373393
#> 93     3   10  1.794387625
#> 94     4   10  0.655421619
#> 95     5   10  0.059149967
#> 96     6   10  0.618199524
#> 97     7   10 -0.653622011
#> 98     8   10  1.251387082
#> 99     9   10  0.308299802
#> 100   10   10  0.051150670