Simulate geostatistical data on set of given locations

Spatially correlated data are simulated assuming a multivariate normal random error vector. For simplicity, only "Exponential" and "Spherical" simulation options are given here.

geostatSim(
  loc.data,
  xcol = "x",
  ycol = "y",
  parsil = 1,
  range = 1,
  nugget = 0,
  minorp = 1,
  rotate = 90,
  extrap = NULL,
  CorModel = "Exponential"
)

Arguments

loc.data

data.frame with x- and y-coordinates of locations for simulated data

xcol

name of the column in loc.data with x-coordinates, default is "x"

ycol

name of the column loc.data with y-coordinates, default is "y"

parsil

partial sill of autocorrelation model, default = 1

range

range of autocorrelation model, default = 1

nugget

range of autocorrelation model, default = 0

minorp

proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1

rotate

rotation of anisotropic axes, default = 90

extrap

extra covariance parameter

CorModel

autocorrelation model, default = "Exponential". Other possibilities are "Spherical".

Value

data.frame of three columns, the original location data appended with a 3rd column of simulated geostatistical data

Author

Jay Ver Hoef

Examples

locations <- expand.grid(1:10, 1:10)
geostatSim(locations, xcol = "Var1", ycol = "Var2",
parsil = 4, range = 20, nugget = 1, CorModel = "Exponential")
#>     Var1 Var2           z
#> 1      1    1  2.92325177
#> 2      2    1  3.26860912
#> 3      3    1  3.27246908
#> 4      4    1  1.55797971
#> 5      5    1  0.14559040
#> 6      6    1 -1.27691270
#> 7      7    1  3.16702606
#> 8      8    1  2.23573566
#> 9      9    1  0.92899385
#> 10    10    1  1.53919656
#> 11     1    2  0.99119948
#> 12     2    2  0.10851680
#> 13     3    2  2.05713777
#> 14     4    2  2.19125343
#> 15     5    2  2.22561094
#> 16     6    2  0.86899395
#> 17     7    2  2.73629090
#> 18     8    2  0.86551692
#> 19     9    2  2.91226399
#> 20    10    2  1.86851777
#> 21     1    3  1.96282993
#> 22     2    3  2.42584801
#> 23     3    3  3.81264415
#> 24     4    3  3.83555733
#> 25     5    3  1.72643866
#> 26     6    3  1.43272253
#> 27     7    3  4.15610442
#> 28     8    3  1.33870756
#> 29     9    3  0.83390060
#> 30    10    3  2.67680735
#> 31     1    4  0.78484506
#> 32     2    4  1.82482052
#> 33     3    4  2.00465140
#> 34     4    4  2.37905990
#> 35     5    4  2.10182925
#> 36     6    4  3.03450030
#> 37     7    4  2.69594423
#> 38     8    4  2.18909831
#> 39     9    4  2.00030880
#> 40    10    4  1.81228676
#> 41     1    5 -0.22492049
#> 42     2    5  0.65929349
#> 43     3    5  2.56924278
#> 44     4    5  2.04929060
#> 45     5    5  3.28342408
#> 46     6    5  5.23079137
#> 47     7    5  2.80817014
#> 48     8    5  2.17458650
#> 49     9    5  2.80671590
#> 50    10    5  2.95262377
#> 51     1    6  3.19900629
#> 52     2    6  1.41683574
#> 53     3    6  4.48556569
#> 54     4    6  3.52431833
#> 55     5    6  5.36378974
#> 56     6    6  3.36584742
#> 57     7    6  4.42059794
#> 58     8    6  3.00384982
#> 59     9    6  3.77898470
#> 60    10    6  1.83887505
#> 61     1    7  3.04543490
#> 62     2    7  2.44828108
#> 63     3    7  2.04427670
#> 64     4    7  1.90682283
#> 65     5    7  2.31110517
#> 66     6    7  2.56777518
#> 67     7    7  3.67509985
#> 68     8    7  2.02867278
#> 69     9    7  0.44307485
#> 70    10    7  1.28011426
#> 71     1    8  2.57126699
#> 72     2    8  1.85021873
#> 73     3    8  3.95748022
#> 74     4    8  3.22143380
#> 75     5    8  3.02455629
#> 76     6    8  2.47675904
#> 77     7    8  0.74592343
#> 78     8    8  3.36309871
#> 79     9    8  2.87527026
#> 80    10    8  0.06673463
#> 81     1    9  4.31828400
#> 82     2    9  1.71652000
#> 83     3    9  3.13095896
#> 84     4    9  3.08335006
#> 85     5    9  3.46217387
#> 86     6    9  3.54836315
#> 87     7    9  4.22818171
#> 88     8    9  3.96712818
#> 89     9    9  2.61022581
#> 90    10    9 -0.73750859
#> 91     1   10  3.77566739
#> 92     2   10  5.92419915
#> 93     3   10  0.54265000
#> 94     4   10  4.21010346
#> 95     5   10  2.16007794
#> 96     6   10  3.48137788
#> 97     7   10  2.82018263
#> 98     8   10  0.85412959
#> 99     9   10  2.21306253
#> 100   10   10  3.64230147