Perform Finite Population Block Kriging

Uses an object of class slmfit from the slmfit() function to predict the response on the unsampled sites. The column of the data set that has the response should have numeric values for the observed response on the sampled sites and `NA` for any site that was not sampled. Note that there is no newdata argument to predict.slmfit: any point in space for which a prediction is needed should be included in the original data set in slmfit() with the response variable as NA.

# S3 method for slmfit
predict(object, wtscol = NULL, ...)

Arguments

object
is an object generated from slmfit()

wtscol
is the name of the column that contains the weights for prediction.

...
further arguments passed to or from other methods. The default setting predicts the population total

Value

a list with

the estimated population total

the estimated prediction variance

a data frame containing

x-coordinates

y-coordinates

density predictions

count predictions

site-by-site density prediction variances

site-by-site count prediction variances

indicator variable for whether or not the each site was sampled

estimated mean for each site

area of each site

vector with estimated covariance parameters

the formula used to fit the model in slmfit()

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predict(slmobj)

#> $FPBK_Prediction
#>          [,1]
#> [1,] 813.1637
#> 
#> $PredVar
#>          [,1]
#> [1,] 607.1478
#> 
#> $Pred_df
#>    counts      pred1       pred2 xcoords ycoords    dummyvar areavar
#> 1      24 0.04819231  0.66770938       1       1 0.836243248     0.5
#> 2      NA 0.07029807 -0.45115262       2       1 0.057692931     1.0
#> 3      21 0.26142376 -0.03514149       3       1 0.438464806     1.0
#> 4      11 0.54271551  0.05692399       4       1 0.906082848     0.5
#> 5      18 0.99896424  1.27839185       5       1 0.203219716     0.5
#> 6      25 0.39053919 -0.78926896       6       1 0.259569137     1.0
#> 7      NA 0.26153323 -1.11039905       7       1 0.002481818     1.0
#> 8      22 0.00985308  0.20476703       1       2 0.839759352     0.5
#> 9      15 0.12668247 -0.71490084       2       2 0.567310578     0.5
#> 10     19 0.67871713 -0.66313734       3       2 0.421974490     1.0
#> 11     16 0.60603457 -0.77168996       4       2 0.739107019     1.0
#> 12     24 0.07165682 -0.25192112       5       2 0.181968057     1.0
#> 13     17 0.89700298 -0.68260187       6       2 0.803160179     1.0
#> 14     17 0.67505385 -0.17863821       7       2 0.349131940     1.0
#> 15     14 0.82949650 -0.02482225       1       3 0.234258575     0.5
#> 16     20 0.10107013 -1.19642613       2       3 0.792952908     1.0
#> 17     14 0.80905060  0.89703610       3       3 0.313615815     0.5
#> 18     28 0.63180551  0.01141215       4       3 0.440336636     0.5
#> 19     NA 0.56321498 -0.79659419       5       3 0.174509722     1.0
#> 20     NA 0.21697451 -0.15645314       6       3 0.541295321     0.5
#> 21     15 0.26280400 -0.08069298       7       3 0.871738130     1.0
#> 22     10 0.23258670  0.18462249       1       4 0.830919171     0.5
#> 23     13 0.91109400 -1.13061351       2       4 0.429160088     0.5
#> 24     NA 0.03107111  0.14803666       3       4 0.282356259     1.0
#> 25     22 0.65086483 -0.25361803       4       4 0.290801075     0.5
#> 26     24 0.99099661  0.61915834       5       4 0.357616148     0.5
#> 27     26 0.21582753 -1.50269581       6       4 0.557304586     1.0
#> 28     11 0.31509692  1.56082594       7       4 0.998539340     1.0
#> 29     30 0.59741644 -1.31280847       1       5 0.340149420     1.0
#> 30     22 0.67413098  0.25418477       2       5 0.572497278     1.0
#> 31     22 0.06157954  0.49486086       3       5 0.727558917     0.5
#> 32     20 0.01452017  1.44691442       4       5 0.196299598     1.0
#> 33     13 0.40946571  2.56768098       5       5 0.928319833     1.0
#> 34     16 0.08069354 -1.80971988       6       5 0.307345031     1.0
#> 35     17 0.54991785  1.18865691       7       5 0.827602321     1.0
#> 36     33 0.29043608  0.74671108       1       6 0.027218879     1.0
#> 37     14 0.52181727  0.37247065       2       6 0.339091076     1.0
#> 38     22 0.42858761 -0.56009555       3       6 0.334321583     1.0
#> 39     21 0.32678279  0.50311496       4       6 0.528220195     0.5
#> 40     23 0.80063605 -0.72036048       5       6 0.271650466     1.0
#>    xcoordsUTM_ ycoordsUTM_ counts_pred_density counts_pred_count
#> 1            1           1            48.00000          24.00000
#> 2            2           1            31.01862          31.01862
#> 3            3           1            21.00000          21.00000
#> 4            4           1            22.00000          11.00000
#> 5            5           1            36.00000          18.00000
#> 6            6           1            25.00000          25.00000
#> 7            7           1            23.02399          23.02399
#> 8            1           2            44.00000          22.00000
#> 9            2           2            30.00000          15.00000
#> 10           3           2            19.00000          19.00000
#> 11           4           2            16.00000          16.00000
#> 12           5           2            24.00000          24.00000
#> 13           6           2            17.00000          17.00000
#> 14           7           2            17.00000          17.00000
#> 15           1           3            28.00000          14.00000
#> 16           2           3            20.00000          20.00000
#> 17           3           3            28.00000          14.00000
#> 18           4           3            56.00000          28.00000
#> 19           5           3            35.51670          35.51670
#> 20           6           3            23.49812          11.74906
#> 21           7           3            15.00000          15.00000
#> 22           1           4            20.00000          10.00000
#> 23           2           4            26.00000          13.00000
#> 24           3           4            32.85532          32.85532
#> 25           4           4            44.00000          22.00000
#> 26           5           4            48.00000          24.00000
#> 27           6           4            26.00000          26.00000
#> 28           7           4            11.00000          11.00000
#> 29           1           5            30.00000          30.00000
#> 30           2           5            22.00000          22.00000
#> 31           3           5            44.00000          22.00000
#> 32           4           5            20.00000          20.00000
#> 33           5           5            13.00000          13.00000
#> 34           6           5            16.00000          16.00000
#> 35           7           5            17.00000          17.00000
#> 36           1           6            33.00000          33.00000
#> 37           2           6            14.00000          14.00000
#> 38           3           6            22.00000          22.00000
#> 39           4           6            42.00000          21.00000
#> 40           5           6            23.00000          23.00000
#>    counts_predvar_density counts_predvar_count counts_sampind counts_muhat
#> 1                  0.0000              0.00000              1     26.34639
#> 2                125.2961            125.29612              0     26.15237
#> 3                  0.0000              0.00000              1     26.63426
#> 4                  0.0000              0.00000              1     27.23214
#> 5                  0.0000              0.00000              1     28.43135
#> 6                  0.0000              0.00000              1     26.73825
#> 7                138.8642            138.86416              0     26.40436
#> 8                  0.0000              0.00000              1     26.16851
#> 9                  0.0000              0.00000              1     26.21182
#> 10                 0.0000              0.00000              1     27.35758
#> 11                 0.0000              0.00000              1     27.18495
#> 12                 0.0000              0.00000              1     26.19780
#> 13                 0.0000              0.00000              1     27.80209
#> 14                 0.0000              0.00000              1     27.45374
#> 15                 0.0000              0.00000              1     27.80411
#> 16                 0.0000              0.00000              1     26.05612
#> 17                 0.0000              0.00000              1     27.95937
#> 18                 0.0000              0.00000              1     27.40552
#> 19               122.2056            122.20560              0     27.09161
#> 20               123.5262             30.88154              0     26.51693
#> 21                 0.0000              0.00000              1     26.62734
#> 22                 0.0000              0.00000              1     26.62202
#> 23                 0.0000              0.00000              1     27.73517
#> 24               126.2106            126.21062              0     26.19998
#> 25                 0.0000              0.00000              1     27.38797
#> 26                 0.0000              0.00000              1     28.27388
#> 27                 0.0000              0.00000              1     26.22645
#> 28                 0.0000              0.00000              1     27.08614
#> 29                 0.0000              0.00000              1     27.05143
#> 30                 0.0000              0.00000              1     27.54447
#> 31                 0.0000              0.00000              1     26.33692
#> 32                 0.0000              0.00000              1     26.44394
#> 33                 0.0000              0.00000              1     27.49559
#> 34                 0.0000              0.00000              1     25.88298
#> 35                 0.0000              0.00000              1     27.48915
#> 36                 0.0000              0.00000              1     26.86122
#> 37                 0.0000              0.00000              1     27.25672
#> 38                 0.0000              0.00000              1     26.86550
#> 39                 0.0000              0.00000              1     26.88379
#> 40                 0.0000              0.00000              1     27.59593
#>    counts_areas
#> 1           0.5
#> 2           1.0
#> 3           1.0
#> 4           0.5
#> 5           0.5
#> 6           1.0
#> 7           1.0
#> 8           0.5
#> 9           0.5
#> 10          1.0
#> 11          1.0
#> 12          1.0
#> 13          1.0
#> 14          1.0
#> 15          0.5
#> 16          1.0
#> 17          0.5
#> 18          0.5
#> 19          1.0
#> 20          0.5
#> 21          1.0
#> 22          0.5
#> 23          0.5
#> 24          1.0
#> 25          0.5
#> 26          0.5
#> 27          1.0
#> 28          1.0
#> 29          1.0
#> 30          1.0
#> 31          0.5
#> 32          1.0
#> 33          1.0
#> 34          1.0
#> 35          1.0
#> 36          1.0
#> 37          1.0
#> 38          1.0
#> 39          0.5
#> 40          1.0
#> 
#> $SpatialParms
#> [1] 2.323211e-04 1.543624e+02 8.240060e-01
#> 
#> $formula
#> counts ~ pred1 + pred2
#> <environment: 0x7fac492d2ac8>
#> 
#> attr(,"class")
#> [1] "sptotalPredOut"