Fits a Spatial Linear Model

Estimates regression coefficients and spatial autocorrelation parameters, given spatial coordinates and a model formula.

slmfit(
  formula,
  data,
  xcoordcol,
  ycoordcol,
  areacol = NULL,
  stratacol = NULL,
  CorModel = "Exponential",
  estmethod = "REML",
  covestimates = c(NA, NA, NA)
)

Arguments

formula

is an R linear model formula specifying the response variable as well as covariates for predicting the response on the unsampled sites.

data

is a data frame or tibble with the response column, the covariates to be used for the block kriging, and the spatial coordinates for all of the sites. Alternatively, data can be an sp Spatial Points Data Frame or sf object with POINT geometry.

xcoordcol

is the name of the column in the data frame with x coordinates or longitudinal coordinates

ycoordcol

is the name of the column in the data frame with y coordinates or latitudinal coordinates

areacol

is the name of the column with the areas of the sites. By default, we assume that all sites have equal area, in which case a vector of 1's is used as the areas.

stratacol

is the name of the the column with the stratification variable, if strata are to be fit separately, with different covariance parameter estimates.

CorModel

is the covariance structure. By default, CorModel is Exponential but other options include the Spherical and Gaussian.

estmethod

is either the default "REML" for restricted maximum likelihood to estimate the covariance parameters and regression coefficients or "ML" to estimate the covariance parameters and regression coefficients. This argument can also be set to "None", in which case covestimates must be provided.

covestimates

is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and estmethod = "None", the the provided vector are treated as the estimators to create the covariance structure.

Value

a list of class slmfit with

the spatial covariance estimates

the regression coefficient estimates

the covariance matrix of the fixed effects

minus two times the log-likelihood of the model

the names of the predictors

the sample size

the name of the covariance model used

a vector of residuals

the design matrix

a vector of the sampled densities

a list containing

formula, the model formula

data, the data set input as the data argument

xcoordcol, the name of the x-coordinate column

ycoordcol, the name of the y-coordinate column

estmethod, either REML or ML

CorModel, the correlation model used

estimated covariance matrix of all sites

Inverted covariance matrix on the sampled sites

the vector of areas.

Examples

data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
summary(slmobj)
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -16.086  -9.120  -4.596   4.963  28.594 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   26.104      4.465   5.847   <2e-16 ***
#> pred1          2.055      6.065   0.339    0.737    
#> pred2          0.214      1.913   0.112    0.912    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            2.323211e-04
#> Partial Sill      1.543624e+02
#> Range             8.240060e-01
#> 
#> Generalized R-squared: 0.00413886 

data(exampledataset) ## load a toy data set
exampledataset$strata <- c(rep("A", 19), rep("B", 21))
strataobj <- slmfit(formula = counts ~ pred1 + pred2,
  data = exampledataset, stratacol = "strata",
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
summary(strataobj)
#> $A
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -11.2360  -7.3855  -0.8668   3.3033  28.1803 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   36.112      4.674   7.726   <2e-16 ***
#> pred1        -13.326      7.892  -1.689   0.1151    
#> pred2         11.161      3.939   2.833   0.0141 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            9.784318e+01
#> Partial Sill      1.534910e-11
#> Range             2.234197e+03
#> 
#> Generalized R-squared: 0.4079598 
#> 
#> $B
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -11.866  -8.444  -4.310   5.431  22.374 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept)   21.754      5.510   3.948  0.00115 **
#> pred1          8.850      9.688   0.914  0.37453   
#> pred2         -1.361      2.349  -0.579  0.57035   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget              0.01384269
#> Partial Sill      142.66630173
#> Range               0.49173852
#> 
#> Generalized R-squared: 0.07575257 
#>