Fits a Separate Spatial Linear Model for Each Stratum

Estimates regression coefficients and spatial autocorrelation parameters, given spatial coordinates, a model formula, and a stratification variable. Arguments are the same here as they are for slmfit(), with an extra argument for stratacol, the name of the stratification column. Note that stratum can either by incorporated as a covariate in slmfit(), in which case the errors have the same spatial covariance, or, models with differing spatial covariances for the errors can be fit to each level of stratum, as is done here in stratafit().

stratafit(
  formula,
  data,
  xcoordcol,
  ycoordcol,
  stratacol = NULL,
  areacol = NULL,
  CorModel = "Exponential",
  estmethod = "REML"
)

Arguments

formula

is an R linear model formula specifying the response variable as well as covariates for predicting the response on the unsampled sites.

data

is the data set with the response column, the covariates to be used for the block kriging, and the spatial coordinates for all of the sites.

xcoordcol

is the name of the column in the data frame with x coordinates or longitudinal coordinates

ycoordcol

is the name of the column in the data frame with y coordinates or latitudinal coordinates

stratacol

is the name of the stratification column

areacol

is the name of the column with the areas of the sites. By default, we assume that all sites have equal area, in which case a vector of 1's is used as the areas.

CorModel

is the covariance structure. By default, CorModel is Exponential but other options include the Spherical and Gaussian.

estmethod

is either the default "REML" for restricted maximum likelihood to estimate the covariance parameters and regression coefficients or "ML" to estimate the covariance parameters and regression coefficients.

Value

a list of class slmfit with

the spatial covariance estimates

the regression coefficient estimates

the covariance matrix of the fixed effects

minus two times the log-likelihood of the model

the names of the predictors

the sample size

the name of the covariance model used

a vector of residuals

the design matrix

a vector of the sampled densities

a list containing

formula, the model formula

data, the data set input as the data argument

xcoordcol, the name of the x-coordinate column

ycoordcol, the name of the y-coordinate column

estmethod, either REML or ML

CorModel, the correlation model used

estimated covariance matrix of all sites

Inverted covariance matrix on the sampled sites

the vector of areas.

Examples

data(exampledataset) ## load a toy data set
exampledataset$strata <- c(rep("A", 19), rep("B", 21))
strataobj <- stratafit(formula = counts ~ pred1 + pred2,
 data = exampledataset, stratacol = "strata",
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
summary(strataobj)
#> $A
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -7.7826 -2.1499  0.0488  1.5492  9.8233 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)  22.2882    35.0930   0.635   0.5364  
#> pred1        -6.5176     3.6384  -1.791   0.0965 .
#> pred2         0.5545     1.7976   0.308   0.7626  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget                 20.0915
#> Partial Sill         1227.2769
#> Range                8575.6681
#> 
#> Generalized R-squared: 0.1998733 
#> 
#> $B
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -9.262 -4.974  1.516  3.355 14.635 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   19.146      2.994   6.395    1e-05 ***
#> pred1          1.430      5.255   0.272    0.789    
#> pred2         -1.603      1.274  -1.259    0.226    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            4.188789e-05
#> Partial Sill      4.205452e+01
#> Range             4.954309e-01
#> 
#> Generalized R-squared: 0.09916644 
#>