Summarizes a fitted spatial linear model with a stratification variable..

If a model is fitted with stratafit(), then this summary function produces summary output for each level of the stratification variable in the same style as the summary.slmfit() function.

# S3 method for class 'stratafit'
summary(object, ...)

Arguments

object

is an object generated from slmfit() of class slmfit.

...

further arguments passed to or from other methods.

Value

a list with

model formula

a table of fixed effects estimates and associated standard errors

estimated spatial covariance parameter estimates

residuals

generalized r-squared.

Examples

data(exampledataset) ## load a toy data set
exampledataset$strata <- c(rep("A", 25), rep("B", 15))
strataobj <- stratafit(formula = counts ~ pred1 + pred2,
 data = exampledataset, stratacol = "strata",
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
summary(strataobj)
#> $A
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -14.033  -7.768  -3.324   4.290  26.857 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   29.846      5.816   5.132    8e-05 ***
#> pred1         -1.223      6.825  -0.179   0.8599    
#> pred2          6.099      3.440   1.773   0.0942 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            1.834005e-03
#> Partial Sill      1.449888e+02
#> Range             1.085866e+00
#> 
#> Generalized R-squared: 0.1571098 
#> 
#> $B
#> 
#> Call:
#> counts ~ pred1 + pred2
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -12.123  -8.845  -4.458   5.275  21.291 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept)   22.895      5.969   3.836  0.00237 **
#> pred1          6.730     11.908   0.565  0.58238   
#> pred2         -1.213      2.683  -0.452  0.65942   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Covariance Parameters:
#>              Exponential Model
#> Nugget            1.531816e+02
#> Partial Sill      1.301003e-11
#> Range             2.638947e+03
#> 
#> Generalized R-squared: 0.04241562 
#>